% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{calc_churn}
\alias{calc_churn}
\title{Calculate churn for a single year (internal salic only - not exported)}
\usage{
calc_churn(x, yr)
}
\arguments{
\item{x}{data frame: Table holding sales by year}

\item{year}{numeric: year to calculate churn (e.g., churn for 2014 uses sales
for 2 years: 2013 and 2014)}
}
\value{
Returns a numeric vector of length 1
}
\description{
This is a simple calculation that doesn't take into account multi-year
licenses, so it should only be used for initial data validation
}
\examples{
library(tidyverse)
library(salic)
load(sale, lic, package = "salic")

y <- select(sale, cust_id, year) \%>\%
    filter(year \%in\% 2012:2013) \%>\% distinct()

calc_churn(y, 2013)
}
\seealso{
Other internal helper functions: \code{\link{format_num}},
  \code{\link{pct_round}}, \code{\link{print_dat}}
}
\keyword{internal}
