% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{carry_duration}
\alias{carry_duration}
\title{Internal Function: Carry multi-year/lifetime durations forward}
\usage{
carry_duration(sale_split, yrs)
}
\arguments{
\item{sale_split}{list: ranked sale table split by year}

\item{yrs}{numeric: Years in sales data (column 'year') from which
to create license history}
}
\description{
This function is intended to be called from \code{\link{make_lic_history}}.
}
\examples{
library(dplyr)
data(sale, lic)

sale_unranked <- left_join(lic, sale)
sale_ranked <- rank_sale(sale_unranked) \%>\%
    select(cust_id, year, duration)

split(sale_ranked, sale_ranked$year) \%>\%     
    carry_duration(2008:2019)
}
\seealso{
Other license history functions: \code{\link{carry_variables}},
  \code{\link{check_history_samp}},
  \code{\link{check_identify_R3}},
  \code{\link{check_identify_lapse}},
  \code{\link{data_join}}, \code{\link{identify_R3}},
  \code{\link{identify_lapse}},
  \code{\link{join_first_month}},
  \code{\link{make_lic_history}}, \code{\link{prep_yrs}},
  \code{\link{rank_sale}}
}
\concept{license history functions}
\keyword{internal}
