% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{cat_caption}
\alias{cat_caption}
\title{Convenience function to print a caption (intended for tables)}
\usage{
cat_caption(caption)
}
\arguments{
\item{caption}{character: Caption to print}
}
\description{
Convenience function to print a caption (intended for tables)
}
\examples{
cat_caption("Here is a caption")
}
\seealso{
Other functions for printing headers, notes, and tables: \code{\link{cat_note}},
  \code{\link{cat_title}}, \code{\link{print_dat}}
}
