% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{cat_note}
\alias{cat_note}
\title{Convenience function to print a note}
\usage{
cat_note(note)
}
\arguments{
\item{note}{character: Note to print}
}
\description{
Convenience function to print a note
}
\examples{
cat_note("here is a note")
}
\seealso{
Other functions for printing headers, notes, and tables: \code{\link{cat_caption}},
  \code{\link{cat_title}}, \code{\link{print_dat}}
}
