% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{cat_title}
\alias{cat_title}
\title{Convenience function to print a title}
\usage{
cat_title(title, note = NULL)
}
\arguments{
\item{title}{character: Title to print}

\item{note}{character: Note to print}
}
\description{
Convenience function to print a title
}
\examples{
cat_title("here is a title")
cat_title("here is a title", "with a note")
}
\seealso{
Other functions for printing headers, notes, and tables: \code{\link{cat_caption}},
  \code{\link{cat_note}}, \code{\link{print_dat}}
}

