% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{catf}
\alias{catf}
\title{Convenience function to print a title}
\usage{
catf(x)
}
\arguments{
\item{x}{character: Contents of title to print}
}
\description{
Convenience function to print a title
}
\examples{
catf()
}
\seealso{
Other functions for for checking license history summarization: \code{\link{printf}}
}
