% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{check_carry_vars}
\alias{check_carry_vars}
\title{Check the carry_vars output of \code{\link{make_lic_history}}}
\usage{
check_carry_vars(lic_history, sale_ranked, carry_vars)
}
\arguments{
\item{lic_history}{license history data frame produced with \code{\link{make_lic_history}}}

\item{sale_ranked}{data frame: Data produced after running
\code{\link{rank_sale}}}

\item{carry_vars}{character: variables to carry over from previous year
(for multi-year and lifetime licenses).}
}
\description{
A test to insure the carry_vars were coded correctly in the license history
table
}
\examples{
# check_carry_vars()
}
\seealso{
Other license history functions: \code{\link{check_history_samp}},
  \code{\link{check_history_summary}},
  \code{\link{check_identify_R3}},
  \code{\link{check_identify_lapse}},
  \code{\link{check_make_priv_final}},
  \code{\link{check_rank_sale}},
  \code{\link{ggplot_lic_history}},
  \code{\link{identify_R3}}, \code{\link{identify_lapse}},
  \code{\link{make_lic_history}}, \code{\link{rank_sale}}
}

