% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_cust}
\alias{check_cust}
\title{Check Customer IDs}
\usage{
check_cust(cust, sale, count = 10000)
}
\arguments{
\item{cust}{data frame: Customer table}

\item{sale}{data frame: Sales table with cust_id & issue_date}

\item{count}{numeric: Number of records to show}
}
\description{
Check Customer IDs
}
\examples{
check_cust()
}
\seealso{
Other functions for checking license data: \code{\link{check_dups}},
  \code{\link{get_dups}}, \code{\link{samp_dups}},
  \code{\link{show_dups}}
}

