% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_dups}
\alias{check_dups}
\title{Check for duplicates in a table}
\usage{
check_dups(x)
}
\arguments{
\item{x}{data frame: Table holding records
intended to speed up the operation}
}
\description{
Note - can probably include a check for duplicate counts, or make another
function for this - see SD prep-cust code
}
\examples{
library(dplyr)
cust <- data.frame(id = 1:3, nm = c("dan k","dan k","dan r"))
x <- select(cust, id, nm) \%>\% distinct()
select(x, nm) \%>\% check_dups()
}
\seealso{
Other functions for checking license data: \code{\link{check_cust}},
  \code{\link{get_dups}}, \code{\link{samp_dups}},
  \code{\link{show_dups}}
}

