% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{check_history_summary}
\alias{check_history_summary}
\title{Check the output of \code{\link{make_lic_history}}}
\usage{
check_history_summary(lic_history)
}
\arguments{
\item{lic_history}{license history data frame produced with \code{\link{make_lic_history}}}
}
\description{
A test to insure that the license history table was built correctly
}
\examples{
# See ?salic::make_lic_history
}
\seealso{
Other license history functions: \code{\link{check_history_samp}},
  \code{\link{check_identify_R3}},
  \code{\link{check_identify_lapse}},
  \code{\link{check_make_priv_final}},
  \code{\link{check_rank_sale}},
  \code{\link{ggplot_lic_history}},
  \code{\link{identify_R3}}, \code{\link{identify_lapse}},
  \code{\link{make_lic_history}}, \code{\link{rank_sale}}
}

