% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{check_identify_lapse}
\alias{check_identify_lapse}
\title{Internal Function: Check lapse identification}
\usage{
check_identify_lapse(lic_history)
}
\arguments{
\item{lic_history}{data frame: output of \code{\link{make_lic_history}}}
}
\description{
Intended to be run as part of \code{\link{identify_lapse}} (where show_summary = TRUE).
Produces a count summary of customers by lapse and lead_year.
}
\examples{
library(dplyr)
data(history)

select(history, -lapse) \%>\%
    filter(year \%in\% 2008:2018) \%>\%
    identify_lapse(2008:2018, show_check_vars = TRUE) \%>\%
    check_identify_lapse()
}
\seealso{
Other license history functions: \code{\link{carry_duration}},
  \code{\link{carry_variables}},
  \code{\link{check_history_samp}},
  \code{\link{check_identify_R3}},
  \code{\link{identify_R3}}, \code{\link{identify_lapse}},
  \code{\link{join_first_month}},
  \code{\link{make_lic_history}}, \code{\link{prep_yrs}},
  \code{\link{rank_sale}}
}
\concept{license history functions}
\keyword{internal}
