% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{check_identify_R3}
\alias{check_identify_R3}
\title{Internal Function: Check R3 Identification}
\usage{
check_identify_R3(lic_history, yrs)
}
\description{
Intended to be run as part of \code{\link{identify_R3}} (where show_summary = TRUE).
Produces a count summary of customers by R3, yrs_since, & lag_duration_run.
}
\examples{
library(dplyr)
data(history)

}
\seealso{
Other deprecated license history functions: \code{\link{check_history_samp}},
  \code{\link{check_identify_lapse}}
}
\concept{deprecated license history functions}
\keyword{internal}
