% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-check.R
\name{check_make_priv_final}
\alias{check_make_priv_final}
\title{Check Final Privilege Table Creation}
\usage{
check_make_priv_final(priv, rnd = 1)
}
\arguments{
\item{priv}{data frame: license privilege tables returned by}

\item{rnd}{numeric: number of decimals to round pct change results
\code{\link{make_priv_final}}}
}
\description{
Check Final Privilege Table Creation
}
\examples{
check_make_priv_final()
}
\seealso{
Other functions for checking license history summarization: \code{\link{check_R3}},
  \code{\link{check_lapse}}, \code{\link{check_make_priv}},
  \code{\link{check_make_track}}, \code{\link{check_track}}
}

