% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-check.R
\name{check_make_track}
\alias{check_make_track}
\title{Check license history creation}
\usage{
check_make_track(track, sale)
}
\arguments{
\item{track}{list: license history tables returned by \code{\link{make_track}}}

\item{sale}{data frame: sales table used to make license history}
}
\description{
Check license history creation
}
\examples{
check_make_track(track, sale_ranked)
}
\seealso{
Other functions for checking license history summarization: \code{\link{check_R3}},
  \code{\link{check_lapse}},
  \code{\link{check_make_priv_final}},
  \code{\link{check_make_priv}}, \code{\link{check_track}}
}

