% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{check_rank_sale}
\alias{check_rank_sale}
\title{Check the output of \code{\link{rank_sale}}}
\usage{
check_rank_sale(sale_ranked, sale_unranked, check_duration = TRUE)
}
\arguments{
\item{sale_ranked}{data frame: Data produced after running
\code{\link{rank_sale}}}

\item{sale_unranked}{data frame: Data used as input to \code{\link{rank_sale}}}

\item{check_duration}{logical: If TRUE, runs a check to see if any
customers have mutliple license sales (with different durations) in the same
year (and insures these are ranked correctly if any are found). If this isn't
a concern, set to FALSE to avoid this relatively time-consuming test.}
}
\description{
A test to insure that sale ranking was performed correctly
}
\examples{
# See ?salic::rank_sale
}
\seealso{
Other license history functions: \code{\link{check_history_samp}},
  \code{\link{check_history_summary}},
  \code{\link{check_identify_R3}},
  \code{\link{check_identify_lapse}},
  \code{\link{check_make_priv_final}},
  \code{\link{ggplot_lic_history}},
  \code{\link{identify_R3}}, \code{\link{identify_lapse}},
  \code{\link{make_lic_history}}, \code{\link{rank_sale}}
}

