% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{check_threshold}
\alias{check_threshold}
\title{Internal Function: Perform an action (e.g., warning, error) to flag records}
\usage{
check_threshold(df, test_threshold, test_variable = "pct_change",
  action = function(...) warning(..., call. = FALSE),
  msg = paste("Threshold of", test_threshold, "for", test_variable,
  "exceeded:"))
}
\arguments{
\item{df}{data frame: table containing statistic to check}

\item{test_threshold}{numeric: if exceeded, while produce warning}

\item{test_variable}{character: Name of variable in df that contains
the test statistic}

\item{action}{function: function call to perform if threshold is exceeded
(intended to be \code{\link[base]{warning}} or \code{\link[base]{stop}}).}

\item{msg}{character: message to be printed in action}
}
\description{
The action is only triggered if any values in df[[test_variable]] exceed
test_threshold. Intended for use in calculating dashboard metrics.
}
\examples{
library(dplyr)

# produce a warning
x <- data.frame(tot = "All", year = 2008:2018, part = rnorm(11, 1000, sd = 100))
x <- mutate(x, pct_change = (part - lag(part)) / lag(part) * 100)
check_threshold(x, 5)

# this will produce an error
# check_threshold(x, 5, action = function(...) stop(..., call. = FALSE))
}
\seealso{
Other dashboard functions: \code{\link{est_churn}},
  \code{\link{est_part}}, \code{\link{format_result}},
  \code{\link{scaleup_part}}
}
\concept{dashboard functions}
\keyword{internal}
