% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history-deprecated.R
\name{check_track}
\alias{check_track}
\title{Check license history (stored in a list)}
\usage{
check_track(x)
}
\arguments{
\item{x}{list: license history tables returned by \code{\link{make_track}}}
}
\description{
Check license history (stored in a list)
}
\examples{
check_track()
}
\seealso{
Other deprecated license history functions: \code{\link{check_R3}},
  \code{\link{check_lapse}}, \code{\link{check_make_priv}},
  \code{\link{check_make_track}}, \code{\link{get_R3}},
  \code{\link{get_lapse}}, \code{\link{make_priv_final}},
  \code{\link{make_priv}}, \code{\link{make_track}},
  \code{\link{update_track}}
}

