% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{cust}
\alias{cust}
\title{Sample Data: 1,000 Hunting/Fishing Customers}
\format{A data frame with 1,000 rows and 4 variables
\describe{
\item{cust_id}{Customer ID (integer) (primary key)}
\item{sex}{Gender (integer) (1=Male, 2=Female)}
\item{birth_year}{Year of Birth (integer) (yyyy)}
\item{cust_res}{State of Residence based on Address (integer) (1=Res, 0=Nonres)}
}}
\description{
Sample Data: 1,000 Hunting/Fishing Customers
}
\seealso{
Other Sample Data: \code{\link{lic}}, \code{\link{sale}},
  \code{\link{state_abbreviations}}
}
\concept{Sample Data}
\keyword{datasets}
