% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{cust_summary_demo}
\alias{cust_summary_demo}
\title{Summarize customers by year for a specified demographic}
\usage{
cust_summary_demo(x, groups)
}
\arguments{
\item{x}{data frame: Table holding sales by year}

\item{groups}{character: grouping variables to be used for the summary}
}
\description{
Note - this needs work to provide more info and the option for csv output
}
\examples{
cust_summary_demo()
}
\seealso{
Other functions for validating license data: \code{\link{calc_churn}},
  \code{\link{lic_summary_revenue}},
  \code{\link{lic_summary}}, \code{\link{pct_round}},
  \code{\link{summary_churn}}, \code{\link{summary_sale}}
}

