% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{dashboard}
\alias{dashboard}
\title{Sample Data: Metrics formatted for Dashboard Input}
\format{A data frame with 7 variables
\describe{
\item{timeframe}{Time period covered ("full-year" or "mid-year")}
\item{group}{Permission group ("all_sports", "hunt", "fish")}
\item{segment}{Demographic group ("All", "Age", "Gender", "Residency")}
\item{year}{Calendar Year (2008, 2009, ..., 2018)}
\item{category}{Category breakouts for segment}
\item{metric}{Metric Summarized ("participants", "recruits", "churn")}
\item{value}{Value of metric for given dimension}
}}
\description{
This data frame holds national/regional dashboard metrics formatted for dashboard input.
The results correspond to full-year metrics for the sample data and include permission 
breakouts for customers aged 18-64 each year. Three permissions are included: 
All Hunters & Anglers (group = "all_sports"), Hunters (group = "hunt"), and 
Anglers (group = "fish") .
}
\seealso{
\code{\link{format_result}}

Other Sample Data: \code{\link{cust}},
  \code{\link{data_salic}}, \code{\link{history}},
  \code{\link{lic}}, \code{\link{metrics}},
  \code{\link{sale}}, \code{\link{state_abbreviations}}
}
\concept{Sample Data}
\keyword{datasets}
