% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{data_check_table}
\alias{data_check_table}
\alias{data_check_lic}
\title{Check format for data tables}
\usage{
data_check_table(df, df_name, required_vars, primary_key, allowed_values,
  no_missing)

data_check_lic(df, df_name = "lic", required_vars = c("lic_id", "type",
  "duration"), allowed_values = list(type = c("fish", "hunt", "combo"),
  duration = 1:99), primary_key = "lic_id", no_missing = "duration")
}
\arguments{
\item{df}{data frame: table to check}

\item{df_name}{character: name of relevant data table (e.g., "lic", "sale", etc.)}

\item{required_vars}{character: variables that should be included}

\item{primary_key}{character: name of variable that acts as primary key (if applicable)}

\item{allowed_values}{list: named list with allowed values for specific variables}

\item{no_missing}{character: name of variables that shouldn't contain missing values}
}
\description{
Any failing checks will produce warnings.
}
\examples{
library(dplyr)
data(lic)

# one table
select(lic, -duration) \%>\%
   data_check_lic()
}
\seealso{
Other functions to check data format: \code{\link{data_required_vars}}
}
\concept{functions to check data format}
\keyword{internal}
