% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-check.R
\name{data_foreign_key}
\alias{data_foreign_key}
\title{Internal Function: Check key coverage between two tables}
\usage{
data_foreign_key(df_foreign, df_primary, key)
}
\arguments{
\item{df_foreign}{data frame: table that stores key as foreign}

\item{df_primary}{data frame: table that stores key as primary}

\item{key}{character: name of variable that acts as key}
}
\description{
Intended to be called from \code{\link{data_check}}, prints a warning if
there are rows in df_foreign with keys that aren't included in df_primary
}
\examples{
library(dplyr)
data(sale, cust)
data_foreign_key(sale, cust, "cust_id")

cust <- filter(cust, cust_id > 5)
data_foreign_key(sale, cust, "cust_id")
}
\seealso{
Other functions to check data format: \code{\link{data_check_table}},
  \code{\link{data_primary_key}},
  \code{\link{variable_allowed_values}}
}
\concept{functions to check data format}
\keyword{internal}
