% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{data_join}
\alias{data_join}
\title{Join all license data (cust, lic, sale)}
\usage{
data_join(cust, lic, sale)
}
\arguments{
\item{cust}{data frame: customer table (key = "cust_id")}

\item{lic}{data frame: license types table (key = "lic_id")}

\item{sale}{data frame: transactions table (foreign keys = "lic_id", "cust_id")}
}
\description{
A convenience fuction to perform inner joins across the 3 tables.
}
\examples{
data(cust, lic, sale)
sale <- data_join(cust, lic, sale)
dplyr::glimpse(sale)
}
\seealso{
Other license history functions: \code{\link{carry_duration}},
  \code{\link{carry_variables}},
  \code{\link{check_history_samp}},
  \code{\link{check_identify_R3}},
  \code{\link{check_identify_lapse}},
  \code{\link{identify_R3}}, \code{\link{identify_lapse}},
  \code{\link{join_first_month}},
  \code{\link{make_lic_history}}, \code{\link{prep_yrs}},
  \code{\link{rank_sale}}
}
\concept{license history functions}
