% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-check.R
\name{data_primary_key}
\alias{data_primary_key}
\alias{data_required_vars}
\alias{data_allowed_values}
\title{Internal Functions: Individual Data Checks}
\usage{
data_primary_key(df, df_name, primary_key)

data_required_vars(df, df_name, required_vars)

data_allowed_values(df, df_name, allowed_values)
}
\arguments{
\item{df}{data frame: table to check}

\item{df_name}{character: name of relevant data table ("cust", "lic", or "sale")}

\item{primary_key}{character: name of variable that acts as primary key,
which should be unique and non-missing. NULL indicates no primary key in table.}

\item{required_vars}{character: variables that should be included}

\item{allowed_values}{list: named list with allowed values for specific variables}
}
\description{
These functions are intended to be called from \code{\link{data_check_table}}. 
Prints a warning message on a failed check.
\itemize{
  \item data_primary_key: check that primary keys are unique and non-missing
  \item data_required_vars: check that required variables are included
  \item data_allowed_values: check that variables are restriced to allowed values.
  Note that \code{\link{data_allowed_values}} is a wrapper for the single variable
  function,  \code{\link{variable_allowed_values}}
}
}
\examples{
library(dplyr)
data(lic)

# primary keys not unique
bind_rows(lic, lic) \%>\% 
    data_primary_key("lic", "lic_id")
    
# primary keys missing
lic$lic_id[1] <- NA
data_primary_key(lic, "lic", "lic_id")

# missing required variables
select(lic, -duration) \%>\%
    data_required_vars("lic", c("lic_id", "type", "duration"))

# includes values that aren't allowed
allowed_values <- list(type = c("hunt", "fish"), duration = 1)
data_allowed_values(lic, "lic", allowed_values)
}
\seealso{
Other functions to check data format: \code{\link{data_check_table}},
  \code{\link{data_check}}, \code{\link{data_foreign_key}},
  \code{\link{variable_allowed_values}}
}
\concept{functions to check data format}
\keyword{internal}
