% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{data_required_vars}
\alias{data_required_vars}
\title{Internal Function: Check for required columns}
\usage{
data_required_vars(df, df_name, required_vars)
}
\arguments{
\item{df}{data frame: table to check}

\item{df_name}{character: name of relevant data table (e.g., "lic", "sale", etc.)}

\item{required_vars}{character: variables that should be included}
}
\description{
Internal Function: Check for required columns
}
\examples{
library(dplyr)
data(lic)

select(lic, -duration, -lic_id) \%>\%
    data_required_vars("lic", c("lic_id", "type", "duration"))
}
\seealso{
Other functions to check data format: \code{\link{data_check_table}}
}
\concept{functions to check data format}
\keyword{internal}
