% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-factor.R
\name{df_factor_var}
\alias{df_factor_var}
\alias{df_factor_sex}
\alias{df_factor_res}
\alias{df_factor_R3}
\alias{df_factor_age}
\title{Convert numeric to factor in data frame and check}
\usage{
df_factor_var(df, var, levels, labels, ...)

df_factor_sex(df, var = "sex", levels = 1:2, labels = c("Male",
  "Female"), ...)

df_factor_res(df, var = "res", levels = c(1, 0),
  labels = c("Resident", "Nonresident"), ...)

df_factor_R3(df, var = "R3", levels = 1:4, labels = c("Carry",
  "Retain", "Reactivate", "Recruit"), ...)

df_factor_age(df, var = "age", levels = 1:7, labels = c("0-17",
  "18-24", "25-34", "35-44", "45-54", "55-64", "65+"), ...)
}
\arguments{
\item{df}{data frame: Input data frame}

\item{var}{numeric: Numeric variable for input}

\item{levels}{numeric: Levels for input numeric vector}

\item{labels}{labels: Labels to use for output factor vector}

\item{...}{Other arguments passed to \code{\link[base]{factor}}}
}
\description{
These are convenience functions for license data, similar to \code{\link{factor_var}}
but operate on data frames (useful for piping) and produce a check summary.
}
\examples{
library(salic)
library(dplyr)
data(cust, sale, package = "salic")
sale2 <- left_join(sale, cust) \%>\% 
    mutate(birth_year = lubridate::year(dob)) \%>\% 
    recode_agecat()
sale2 <- sale2 \%>\%
    select(-agecat) \%>\%
    df_factor_age()
}
\seealso{
Other functions for labelling numeric variables: \code{\link{factor_var}}
}
\concept{functions for labelling numeric variables}
