% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{est_churn}
\alias{est_churn}
\title{Estimate churn by year from license history}
\usage{
est_churn(history, segment = "tot", test_threshold = 30,
  show_test_stat = FALSE, suppress_warning = FALSE)
}
\arguments{
\item{history}{data frame: input license history table}

\item{segment}{character: defaults to "tot", which indicates no segmentation.
Alternatively specifiy other license history variables ("res", "sex", etc.)}

\item{test_threshold}{numeric: threshold in whole number percentage points 
for pct change per year. A warning will be printed if the absolute value
of the change for any year exceeds the threshold.}

\item{show_test_stat}{logical: If TRUE, the output table will include
a variable holding the test statistic for each row.}

\item{suppress_warning}{logical: If TRUE, no test warning will be displayed 
(even if threshold is exceeded). Test statistics can still be included by 
setting show_test_stat = TRUE.}
}
\description{
This function requires a correctly formated history table (see \code{\link{history}}).
It runs a mean of the lapse value (per year), optionally by segment (and also shifts 
year forward by 1 so that churn in current year reflects lapse pct from last year). 
It also runs a validation test: pct change per year.
}
\examples{
library(dplyr)
data(history)
history <- history \%>\%
    label_categories() \%>\%
    recode_agecat() \%>\%
    filter(!agecat \%in\% c("0-17", "65+"), year != 2019)
est_churn(history)

# apply across all segments
segs <- c("tot", "res", "sex", "agecat")
churn <- sapply(segs, function(x) est_churn(history, x), simplify = FALSE)
}
\seealso{
Other dashboard functions: \code{\link{check_threshold}},
  \code{\link{est_part}}, \code{\link{scaleup_part}}
}
\concept{dashboard functions}
