% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{est_recruit}
\alias{est_recruit}
\title{Estimate new participants by year from license history}
\usage{
est_recruit(history, segment = "tot", test_threshold = 30,
  show_test_stat = FALSE, suppress_warning = FALSE,
  outvar = "recruit", recruit_values = c("Recruit", "4"))
}
\arguments{
\item{history}{data frame: input license history table}

\item{segment}{character: defaults to "tot", which indicates no segmentation.
Alternatively specifiy other license history variables ("res", "sex", etc.)}

\item{test_threshold}{numeric: threshold in whole number percentage points 
for pct change per year. A warning will be printed if the absolute value
of the change for any year exceeds the threshold.}

\item{show_test_stat}{logical: If TRUE, the output table will include
a variable holding the test statistic for each row.}

\item{suppress_warning}{logical: If TRUE, no test warning will be displayed 
(even if threshold is exceeded). Test statistics can still be included by 
setting show_test_stat = TRUE.}

\item{outvar}{character: name of variable that stores metric}

\item{recruit_values}{character: Values of R3 variable that indicate a recruit
(passed to \code{\link[dplyr]{filter}} to exclude non-recruits)}
}
\description{
This function requires a correctly formated history table (see \code{\link{history}}).
It produces a simple count of records per year (filtered by R3 to exclude non-recruits),
optionally by segment. It also runs a validation test: pct change per year.
}
\examples{
library(dplyr)
data(history)

history <- filter(history, year != 2019)
est_recruit(history)

# works whether or not R3 variable has been labelled
count(history, R3)
history <- label_categories(history)
count(history, R3)
est_recruit(history)
}
\seealso{
Other dashboard functions: \code{\link{check_threshold}},
  \code{\link{est_churn}}, \code{\link{est_part}},
  \code{\link{scaleup_part}}
}
\concept{dashboard functions}
