% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category-variables.R
\name{factor_var}
\alias{factor_var}
\alias{factor_age}
\alias{factor_sex}
\alias{factor_res}
\alias{factor_R3}
\title{Convert numeric variable to factor}
\usage{
factor_var(x, levels, labels, suppress_check = TRUE, ...)

factor_age(x, levels = 1:7, labels = c("0-17", "18-24", "25-34",
  "35-44", "45-54", "55-64", "65+"), ...)

factor_sex(x, levels = 1:2, labels = c("Male", "Female"), ...)

factor_res(x, levels = c(1, 0), labels = c("Resident", "Nonresident"),
  ...)

factor_R3(x, levels = 1:4, labels = c("Carry", "Renew", "Reactivate",
  "Recruit"), ...)
}
\arguments{
\item{x}{numeric: Input numeric vector}

\item{levels}{numeric: Levels for input numeric vector}

\item{labels}{labels: Labels to use for output factor vector}

\item{suppress_check}{logical: If TRUE, does not print a coding summary}

\item{...}{Other arguments passed to \code{\link[base]{factor}}}
}
\description{
These are convenience functions for use with license data. They are
wrappers for \code{\link[base]{factor}} with specified values.
}
\examples{
library(dplyr)
data(history)
x <- history \%>\% mutate(
    R3 = factor_R3(R3, suppress_check = FALSE),
    sex = factor_sex(sex, suppress_check = FALSE),
    res = factor_res(res, suppress_check = FALSE)
)
}
\seealso{
Other functions for working with category variables: \code{\link{df_factor_var}},
  \code{\link{label_categories}},
  \code{\link{recode_agecat}}
}
\concept{functions for working with category variables}
