% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{format_num}
\alias{format_num}
\title{Format numbers for printing (internal salic only - not exported)}
\usage{
format_num(x, digits = 1, big.mark = ",")
}
\arguments{
\item{x}{vector to format}

\item{digits}{number of digits for rounding}

\item{big.mark}{character: separator between 1000s}
}
\description{
Helper function for use in print_dat
}
\examples{
format_num(c(2005, 2006, NA))
format_num(c(100000, 131000, 150000, NA))
format_num(c(100000, 131000, 150000, NA), big.mark = "")
format_num(c(27.456, 25.234, 30.679, NA))
format_num(c(27, 25, 30))
format_num(c("6.2\%", "13.9\%", "8.2\%", NA))
}
\seealso{
Other internal helper functions: \code{\link{calc_churn}},
  \code{\link{pct_round}}, \code{\link{print_dat}}
}
\concept{internal helper functions}
\keyword{internal}
