% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{format_tableau}
\alias{format_tableau}
\title{Format estimated metrics for input to Tableau Dashboard}
\usage{
format_tableau(df, timeframe, group)
}
\arguments{
\item{df}{data frame: Input table with estimated metrics}

\item{timeframe}{character: value to store in the 'timeframe' variable of 
the output (e.g, 'full-year', 'mid-year')}

\item{group}{character: value to sore in the 'group' variable of the 
output (e.g., 'all_sports', 'fish', 'hunt')}
}
\description{
This function expects a data frame produced by an salic estimation function (
\code{\link{est_part}}, \code{\link{est_recruit}}, \code{\link{est_churn}}).
It returns a data frame with additional formatting that allows stacking all results
into a single table.
}
\examples{
library(dplyr)
data(metrics)

# format some different tables
x <- format_tableau(metrics$participants$tot, "full-year", "all_sports")
y <- format_tableau(metrics$participants$res, "full-year", "all_sports")
z <- format_tableau(metrics$recruits$sex, "full-year", "all_sports")
bind_rows(x, y, z) # combine

# apply formatting across segments
lapply(metrics$participants, function(x) format_tableau(x, "full-year", "sports")) \%>\%
   bind_rows
   
# apply across metrics & segments
bind_rows(
    lapply(metrics$participants, function(x) format_tableau(x, "full-year", "sports")),
    lapply(metrics$recruits, function(x) format_tableau(x, "full-year", "sports")),
    lapply(metrics$churn, function(x) format_tableau(x, "full-year", "sports"))
)
}
\seealso{
Other dashboard functions: \code{\link{check_threshold}},
  \code{\link{est_churn}}, \code{\link{est_part}},
  \code{\link{scaleup_part}}
}
\concept{dashboard functions}
