% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{get_dups}
\alias{get_dups}
\title{Get duplicates based on an ID}
\usage{
get_dups(x, full, id)
}
\arguments{
\item{x}{data frame: Table holding distinct observations}

\item{full}{data frame: Table holding all observations}

\item{id}{character: Variable(s) representing identifier to check}
}
\description{
Note - can probably remove "full" parameter
}
\examples{
library(dplyr)
cust <- data.frame(id = 100001:100003, nm = c("dan k","dan k","dan r"))
x <- select(cust, id, nm) \%>\% distinct()
select(x, nm) \%>\% check_dups()
get_dups(x, cust, "nm")
}
\seealso{
Other functions for checking license data: \code{\link{check_cust}},
  \code{\link{check_dups}}, \code{\link{samp_dups}},
  \code{\link{show_dups}}
}

