% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{get_lapse}
\alias{get_lapse}
\title{Identify lapse group each year}
\usage{
get_lapse(track, i, count_multi = TRUE)
}
\arguments{
\item{track}{list: Set of license history tables created using
\code{\link{make_track}}}

\item{i}{integer/character: position of data frame in tracking list to
calculate metric}

\item{count_multi}{logical: If TRUE, carried group (multi-year, lifetime) is
counted as renewed for calculating lapse (churn)}
}
\description{
This is called as part of \code{\link{make_priv}} and codes lapse:
0 = renew next year, 1 = lapse next year.
}
\details{
Lapse Conditionals:
\itemize{
\item \emph{1 (lapse)}: Didn't buy a license next year
(duration == 0 & left == 0)
\item \emph{0 (renew)}: Did buy a license next year
(duration >= 1 & left == 0)
\item \emph{0 (carry)}: Multi-year or lifetime license next year
(left > 0 & !is.na(left))
}
}
\examples{
# Calculate lapse in 2012
library(dplyr)
x <- get_lapse(track, "2012")
count(x, duration, left, lapse) \%>\%
    filter(!is.na(R3)) \%>\%
    tidyr::spread(duration, n, fill = "") \%>\% data.frame()
}
\seealso{
Other functions for summarizing license history: \code{\link{get_R3}},
  \code{\link{make_priv_final}}, \code{\link{make_priv}},
  \code{\link{make_track}}, \code{\link{rank_sale}},
  \code{\link{update_track}}
}

