% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{ggplot_lic_history}
\alias{ggplot_lic_history}
\title{Check the output of \code{\link{make_lic_history}} using a ggplot geom_bar}
\usage{
ggplot_lic_history(lic_history, yrs, var = "duration_run")
}
\arguments{
\item{lic_history}{license history data frame produced with \code{\link{make_lic_history}}}

\item{yrs}{numeric: Years in sales data (column 'year') from which
to create license history}

\item{var}{character: Name of variable to use for "fill" parameter in 
\code{\link[ggplot2]{aes}} in \code{\link[ggplot2]{geom_bar}}}
}
\description{
A visual check of license history for a given variable (duration_run, R3, lapse)
}
\examples{
# See ?salic::make_lic_history
}
\seealso{
Other license history functions: \code{\link{check_history_samp}},
  \code{\link{check_history_summary}},
  \code{\link{check_identify_R3}},
  \code{\link{check_identify_lapse}},
  \code{\link{check_make_priv_final}},
  \code{\link{check_rank_sale}}, \code{\link{identify_R3}},
  \code{\link{identify_lapse}},
  \code{\link{make_lic_history}}, \code{\link{rank_sale}}
}

