% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{history}
\alias{history}
\title{Sample Data: License  History}
\format{A data frame with 9 variables
\describe{
\item{cust_id}{Customer ID (integer) (composite key)}
\item{year}{License Year (integer) (yyyy) (composite key)}
\item{duration_run}{Running Duration in years (integer)}
\item{month}{Earliest Transaction Month (integer) (1=Jan, 2=Feb, ...)}
\item{res}{Residency (integer) (1=Res, 0=Nonres)}
\item{sex}{Gender (integer) (1=Male, 2=Female)}
\item{birth_year}{Year of Birth (integer) (yyyy)}
\item{R3}{R3 Group this Year (integer) (1=carry, 2=renew, 3=reactivate, 4=recruit)}
\item{lapse}{Lapsed next Year? (integer) (1=lapse, 0=renew)}
}}
\description{
This is a table that corresponds to the "all_sports" permission for the included
license sample data; containing 1 row per customer per year a license was held.
}
\seealso{
\code{\link{rank_sale}} \code{\link{make_history}}

Other Sample Data: \code{\link{cust}},
  \code{\link{dashboard}}, \code{\link{lic}},
  \code{\link{metrics}}, \code{\link{sale}},
  \code{\link{state_abbreviations}}
}
\concept{Sample Data}
\keyword{datasets}
