% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{forward_duration}
\alias{forward_duration}
\alias{history_internal}
\title{Internal Functions: Making license history}
\usage{
forward_duration(df, df_last, current_year, carry_vars)
}
\arguments{
\item{df}{data frame: current year table}

\item{df_last}{data frame: previous year table}

\item{carry_vars}{character: additional variables to carry over from previous year
(for multi-year and lifetime licenses).}
}
\description{
Intended to be called from \code{\link{make_history}}.
}
\examples{
# example
}
\seealso{
Other license history functions: \code{\link{join_first_month}},
  \code{\link{make_history}}, \code{\link{prep_yrs}},
  \code{\link{rank_sale}}
}
\concept{license history functions}
\keyword{internal}
