% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{identify_lapse}
\alias{identify_lapse}
\title{Identify lapse group each year}
\usage{
identify_lapse(lic_history, yrs, show_summary = FALSE,
  show_check_vars = FALSE)
}
\arguments{
\item{lic_history}{license history data frame produced with \code{\link{make_lic_history}}}

\item{yrs}{numeric: Years in sales data (column 'year') from which
to create license history}

\item{show_summary}{logical: if TRUE, include a tabular summary for checking (by running 
\code{\link{check_identify_R3}})}

\item{show_check_vars}{logical: if TRUE, include output variables used in summary 
(lag_duration_run, yrs_since)}
}
\description{
This is intended to be called following \code{\link{make_lic_history}} 
and codes lapse (0 = renews next year, 1 = lapses next year).
}
\examples{
library(dplyr)
data(history)

yrs <- 2008:2019
x <- history \%>\%
    select(-lapse) \%>\%
    # 2019 is incomplete (1st 6 months only) & must be excluded from the function call
    identify_lapse(yrs[-length(yrs)], show_summary = TRUE)

# calculate churn rate
group_by(x, year) \%>\% 
    summarise(mean(lapse)) \%>\%
    mutate(year = year + 1)
}
\seealso{
Other license history functions: \code{\link{check_history_samp}},
  \code{\link{check_identify_R3}},
  \code{\link{check_identify_lapse}},
  \code{\link{identify_R3}},
  \code{\link{join_first_month}},
  \code{\link{make_lic_history}}, \code{\link{rank_sale}}
}
\concept{license history functions}
