% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category-variables.R
\name{label_categories}
\alias{label_categories}
\title{Convert multiple numeric variables to factors}
\usage{
label_categories(df, categories = c("R3", "sex", "res"), ...)
}
\arguments{
\item{df}{data frame: Input data frame}

\item{categories}{character: vector of variable names to convert to factor
(if present)}

\item{...}{additional arguments passed to \code{\link{df_factor_var}}}
}
\description{
This is a convenience function to convert the expected set of numeric
category variables to factors. It's basically a wrapper for
 \code{\link{df_factor_var}}.
}
\examples{
library(dplyr)
data(history)
x <- label_categories(history, suppress_check = FALSE)
}
\seealso{
Salic Function Reference: \code{\link{salic}}

Other functions for working with category variables: \code{\link{df_factor_var}},
  \code{\link{factor_var}}, \code{\link{recode_agecat}}
}
\concept{functions for working with category variables}
