% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{lic}
\alias{lic}
\title{Sample Data: License Types}
\format{A data frame with 3 variables
\describe{
\item{lic_id}{License ID (numeric) (primary key) - Note that IDs can be stored 
as character, but this is less efficient.}
\item{type}{License Type (character) ("hunt" = hunting license, "fish" = 
fishing license,  "combo" = combination hunting/fishing license)}
\item{duration}{Duration in years (numeric) (1 = 1yr/short-term, 2 = 2yr,..., 
99 = lifetime)}
}}
\description{
Sample Data: License Types
}
\seealso{
Other Sample Data: \code{\link{cust}},
  \code{\link{dashboard}}, \code{\link{data_salic}},
  \code{\link{history}}, \code{\link{metrics}},
  \code{\link{sale}}, \code{\link{state_abbreviations}}
}
\concept{Sample Data}
\keyword{datasets}
