% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{lic_history}
\alias{lic_history}
\title{Sample Data: License  History (2008 through 2019)}
\format{A data frame with 9,672 rows and 8 variables
\describe{
\item{cust_id}{Customer ID (integer) (composite key)}
\item{year}{License Year (integer) (yyyy) (composite key)}
\item{month}{Earliest Transaction Month (integer) (1=Jan, 2=Feb, ..., 12=Dec)}
\item{lapse}{Lapsed next Year? (integer) (1=lapse, 0=renew)}
\item{R3}{R3 Group this Year (integer) (1=carry, 2=renew, 3=reactivate, 4=recruit)}
\item{res}{Residency (integer) (1=Res, 0=Nonres)}
\item{sex}{Gender (integer) (1=Male, 2=Female)}
\item{birth_year}{Year of Birth (integer) (yyyy)}
}}
\description{
This is a table that corresponds to the "all_sports" permission for the included
license sample data; containing 1 row per customer per year a license was held.
Only the variables needed for dashboard production were included (i.e., certain 
variables potentially useful for validation, such as duration_run, were removed).
}
\seealso{
\code{\link{make_lic_history}}

Other Sample Data: \code{\link{cust}}, \code{\link{lic}},
  \code{\link{sale}}, \code{\link{state_abbreviations}}
}
\concept{Sample Data}
\keyword{datasets}
