% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{lic_summary_revenue}
\alias{lic_summary_revenue}
\title{Summarize license revenue by selected grouping variables}
\usage{
lic_summary_revenue(x, groups, show_change = FALSE, out = NULL,
  title = "Revenue Sum by Year", note = NULL)
}
\arguments{
\item{x}{data frame: Table holding sales by year}

\item{groups}{character: grouping variables to be used for the summary}

\item{show_change}{logical: If TRUE, include a table showing percent changes
by year}

\item{out}{character: file path to optional output csv}

\item{title}{character: title for output table}

\item{note}{character: note for output table}
}
\description{
Summarize license revenue by selected grouping variables
}
\examples{
library(tidyverse)
library(salic)
load(sale, lic, package = "salic")

x <- select(lic, lic_id, description) \%>\% left_join(sale) \%>\% mutate(revenue = 30)

lic_summary_revenue(x, c("lic_id", "description"))
lic_summary_revenue(x, c("lic_id", "description"), out = "lic_summary_rev.csv")

lic_summary_revenue(x, c("lic_id", "description"), title = "Title", note = "Note")
}
\seealso{
Other functions for validating license data: \code{\link{calc_churn}},
  \code{\link{count_lines_textfile}},
  \code{\link{cust_summary_demo}},
  \code{\link{lic_summary}}, \code{\link{pct_round}},
  \code{\link{summary_churn}}, \code{\link{summary_sale}}
}
