% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{make_history}
\alias{make_history}
\title{Make license history table}
\usage{
make_history(sale_ranked, yrs, carry_vars = NULL, yrs_lapse = yrs,
  show_diagnostics = FALSE)
}
\arguments{
\item{sale_ranked}{data frame: Sales table from which license history will be made}

\item{yrs}{numeric: Years in sales data (column 'year') from which
to create license history}

\item{carry_vars}{character: additional variables to carry over from previous year
(for multi-year and lifetime licenses).}

\item{yrs_lapse}{numeric: years to include in lapse calculation (defaults to yrs). 
If NULL, lapse will not be calculated (useful for mid-year results)}

\item{show_diagnostics:}{If TRUE, will include intermediate variables in the
output dataset, useful for running checks.}
}
\description{
The license history table accounts for multi-year/lifetime licenses directly by including
a row for every year a license is held. The input data frame (sale_ranked) must 
have at least 3 columns (cust_id, year, duration) and should only have 1 record 
per customer per year (ensured by running \code{\link{rank_sale}} beforehand).
The output data frame includes the following variables:
\itemize{
\item \emph{cust_id}: Customer ID
\item \emph{year}: License Year
\item \emph{carry_vars}: One or more optional variables to include; ensures 
multi-year/lifetime records are complete in future years.
\item \emph{duration_run}: A duration that accounts for multi-year/lifetimes.
\item \emph{lapse}: Lapse next year? (1=Yes, 0=No). Only included if yrs_lapse != NULL
\item \emph{R3}: R3 group this year (1=carried, 2=renewed, 3=reactivated, 4=recruited).  
Only included if more than 5 years are present.
}
Development Note: This function makes use of several \code{\link{history_internal}}
functions.
}
\examples{
library(dplyr)
data(sale, lic)
sale_ranked <- left_join(sale, lic) \%>\% rank_sale()
history <- make_history(sale_ranked, 2008:2018, "res")

# history includes more rows than sale_ranked if multi-year/lifetimes are present
left_join(count(history, year), count(sale_ranked, year), by = "year")

# run with diagnostics for checking
# TODO
}
\seealso{
Other license history functions: \code{\link{join_first_month}},
  \code{\link{prep_yrs}}, \code{\link{rank_sale}}
}
\concept{license history functions}
