% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{make_lic_history}
\alias{make_lic_history}
\title{Make a License History table with 1 row per customer per year}
\usage{
make_lic_history(sale_ranked, yrs, carry_vars = NULL)
}
\arguments{
\item{sale_ranked}{data frame: Sales table from which license history will be made}

\item{yrs}{numeric: Years in sales data (column 'year') from which
to create license history}

\item{carry_vars}{character: additional variables to carry over from previous year
(for multi-year and lifetime licenses).}
}
\description{
The license history table accounts for multi-year/lifetime licenses directly by including
a row for every year a license is held. The input data frame (sale_ranked) must 
have at least 3 columns (cust_id, year, duration) and should only have 1 record 
per customer per year (ensured by running \code{\link{rank_sale}} beforehand).
}
\details{
The output data frame includes the following variables:
\itemize{
\item \emph{cust_id}
\item \emph{year}
\item \emph{duration_run}: A duration variable that accounts for multi-year/lifetimes, 
also used in downstream R3 calculations.
\item \emph{carry_vars}: One or more optional variables to include, 
based on the corresponding argument (typically carry_vars = c("month", "res")).
For multi-year/lifetimes, carry_vars inherit values from the previous year.
}
Development Note: This function is largely a wrapper for \code{\link{carry_duration}} 
and \code{\link{carry_variables}}
}
\examples{
library(dplyr)
data(lic, sale)
sale_unranked <- left_join(sale, lic)
sale_ranked <- rank_sale(sale_unranked) \%>\%
    join_first_month(sale_unranked)
history <- sale_ranked \%>\%
    make_lic_history(2008:2019, carry_vars = c("month", "res"))

# check a sample of several customers
check_history_samp(history)
}
\seealso{
Other license history functions: \code{\link{carry_duration}},
  \code{\link{carry_variables}},
  \code{\link{check_history_samp}},
  \code{\link{check_identify_R3}},
  \code{\link{check_identify_lapse}},
  \code{\link{data_join}}, \code{\link{identify_R3}},
  \code{\link{identify_lapse}},
  \code{\link{join_first_month}}, \code{\link{prep_yrs}},
  \code{\link{rank_sale}}
}
\concept{license history functions}
