% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{make_priv_final}
\alias{make_priv_final}
\title{Finalizes input privilege table (stored as list)}
\usage{
make_priv_final(priv, keep = c("cust_id", "year", "lapse", "R3", "res", "sex",
  "age", "county"))
}
\arguments{
\item{priv}{list: Input privilege table}

\item{keep}{character: Variables to keep in the cleaned table}
}
\description{
This returns a final privilege table - stored as a data frame. Drops rows
that don't represent license privileges & finalizes variables
}
\details{
The output data frame has 4 columns:
\itemize{
\item \emph{cust_id}: Customer ID
\item \emph{year}: License Year
\item \emph{lapse}: Lapsed next year? (==1: yes, ==0: no)
\item \emph{R3}: R3 group this year (Carried, Renewed, Reactivated, Recruited)
}
}
\examples{
make_priv_final()
}
\seealso{
Other functions for summarizing license history: \code{\link{get_R3}},
  \code{\link{get_lapse}}, \code{\link{make_priv}},
  \code{\link{make_track}}, \code{\link{rank_sale}},
  \code{\link{update_track}}
}

