% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{make_track}
\alias{make_track}
\title{Make a License History Table}
\usage{
make_track(sale_ranked, yrs, carry_vars = NULL)
}
\arguments{
\item{sale_ranked}{data frame: Sales table from which license history will be made}

\item{yrs}{numeric: Years in sales data (column 'year') from which
to create license history}

\item{carry_vars}{character: variables to carry over from previous year
(for multi-year and lifetime licenses)}
}
\description{
Output table is stored in a list of data frames (one per license year). It is
intended  as intermediate output for making a privilege table using
\code{\link{make_priv}}.
}
\details{
The input data frame (sale) must have at least 3 columns
(cust_id, year, duration) and should only have 1 record per customer per year
(which can be insured by running \code{\link{rank_sale}}).

Each data frame in the output list has 3 columns:
\itemize{
\item \emph{cust_id}: Customer ID
\item \emph{duration}: Bought a license this year? (>=1: yes, ==0: no),
where duration = 1 refers to annual or short-term, 2 to 2 yr, 3 to 3 yr, ...,
duration = 99 is used to indicate lifetime licenses
\item \emph{left}: Time remaining on previous license  (==0: need to renew,
>0: don't need to renew, ==NA: haven't bought before, <0 didn't buy last year)
}
}
\examples{
# Make tracking table using sample data (and check)
# this makes a tracking table that matches the one pre-loaded in sample data
track <- make_track(sale_ranked, 2004:2013)
check_make_track(track, sale_ranked)
}
\seealso{
Other functions for summarizing license history: \code{\link{get_R3}},
  \code{\link{get_lapse}}, \code{\link{make_priv_final}},
  \code{\link{make_priv}}, \code{\link{rank_sale}},
  \code{\link{update_track}}
}

