% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{metrics}
\alias{metrics}
\title{Sample Data: Dashboard Metrics}
\format{A list of 3 metrics (participants, recruits, churn) with 4 segments each 
(tot, res, sex, agecat): 12 data frames total
\describe{
\item{participants}{Sportspersons per Year}
\item{recruits}{New Sportspersons (i.e., recruits) per Year}
\item{churn}{Churn (i.e., turnover) per Year}
}}
\description{
This list holds national/regional dashboard metrics (full-year) for the sample 
data and corresponds to "All Hunters & Anglers" aged 18-64 each year. Each element 
(metric) holds four segments ("tot" = overall, "res" = by Residency, "sex" = by Gender, 
"agecat" = by Age Category)
}
\seealso{
Salic Function Reference: \code{\link{salic}}

Calculating Metrics: \code{\link{est_part}}, \code{\link{est_churn}}

Other Sample Data: \code{\link{cust}},
  \code{\link{dashboard}}, \code{\link{data_salic}},
  \code{\link{history}}, \code{\link{lic}},
  \code{\link{sale}}
}
\concept{Sample Data}
\keyword{datasets}
