% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{new_section_yr1}
\alias{new_section_yr1}
\title{Create a new analysis data sub-directory with template scripts}
\usage{
new_section_yr1(num, cust_file = TRUE, projdir = NULL)
}
\arguments{
\item{num}{integer: Number for script order (1 >> prep-state, etc.)}

\item{cust_file}{logical: If TRUE, template scripts assume separate files for
customer and transaction data (otherwise they are assumed to be stored together)}

\item{projdir}{character: Subdirectory to place folders/files. If NULL,
the working directory will be used}
}
\description{
This is a variation on saproj \code{\link[saproj]{new_section}}. Section titles
and template scripts are populated given the selected number
}
\examples{
saproj::new_project(projdir = "test-dir")
salic::new_section_yr1(1, projdir = "test-dir")
salic::section(2, projdir = "test-dir")
}
