% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pct_round}
\alias{pct_round}
\title{Round numeric values and print as a character percentages}
\usage{
pct_round(x, rnd = 1, scale = 100)
}
\arguments{
\item{x}{numeric: Vector of values to display as rounded percentages}

\item{rnd}{numeric: number of decimals to round pct change results}

\item{scale}{numeric: scaling paramter - defaults to 100 for showing percentages}
}
\description{
This is a helper function for conveniently displaying percentages. It is used
in other summary functions in salic.
}
\examples{
x <- data.frame(id = c(1,2,3,4), pop = c(135, 416, 389, 320))
x$pct <- pct_round(x$pop / sum(x$pop))
print_dat(x, "Population Percentages")
}
\seealso{
Other functions for validating license data: \code{\link{calc_churn}},
  \code{\link{count_lines_textfile}},
  \code{\link{cust_summary_demo}},
  \code{\link{lic_summary_revenue}},
  \code{\link{lic_summary}}, \code{\link{summary_churn}},
  \code{\link{summary_sale}}
}
