% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{prep_yrs}
\alias{prep_yrs}
\title{Internal Function: Check years range & sort}
\usage{
prep_yrs(yrs, df, func_name)
}
\arguments{
\item{yrs}{numeric: Years in sales data (column 'year') from which
to create license history}

\item{df}{data frame: table that contains "year" variable}

\item{func_name}{character: name of function to print in warning}
}
\description{
Prints a warning any of the specified years aren't included in the dataset and 
returns a vector of availabe years, sorted ascending. The sort ensures correct 
ordering in subsequent license history calculations, which include iterations
by year that would produce incorrect results if not sorted.
}
\details{
This function is intended to be called from \code{\link{make_history}}
}
\examples{
data(sale)
prep_yrs(c(2010, 2008, 2015), sale, "my_function")
prep_yrs(c(2007, 2015, 2010), sale, "my_function") # print a warning
}
\seealso{
Other internal license history functions: \code{\link{history_internal}}
}
\concept{internal license history functions}
\keyword{internal}
