% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{printf}
\alias{printf}
\title{Convenience function to print contents of a data frame}
\usage{
printf(x)
}
\arguments{
\item{x}{data frame: data frame contents to print}
}
\description{
Convenience function to print contents of a data frame
}
\examples{
printf()
}
\seealso{
Other functions for for checking license history summarization: \code{\link{catf}},
  \code{\link{check_rank_sale}}
}

