% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{rank_month}
\alias{rank_month}
\title{Join earliest sale month by customer-year to ranked sale table.}
\usage{
rank_month(sale_ranked, sale_unranked)
}
\arguments{
\item{sale_ranked}{data frame: Input ranked sales data}

\item{sale_unranked}{data frame: Input unranked sales data}
}
\description{
This function is only intended to be run following \code{\link{rank_sale}}
}
\examples{
### Run ranking with sample data
library(tidyverse)
data(lic, sale, package = "salic")

# Join duration variable to sale data
sale_unranked <- select(lic, lic_id, duration) \%>\%
    right_join(sale) \%>\%
    select(cust_id, year, duration)
    
# Randomly generate a month variable for this example
sale_unranked$month <- base::sample(1:12, nrow(sale_unranked), replace = T)

# Perform ranking
sale_ranked <- sale_unranked \%>\%
    rank_sale() \%>\%
    rank_month(sale_unranked)

# check month ranking
left_join(
    count(sale_ranked, month), 
    distinct(sale_unranked, cust_id, year, month) \%>\% count(month), 
    by = "month"
)
}
\seealso{
Other license history functions: \code{\link{check_history_samp}},
  \code{\link{check_identify_R3}},
  \code{\link{check_identify_lapse}},
  \code{\link{identify_R3}}, \code{\link{identify_lapse}},
  \code{\link{make_lic_history}}, \code{\link{rank_sale}}
}
\concept{license history functions}
