% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{rank_sale}
\alias{rank_sale}
\title{Filter sales to 1 row per customer per year.}
\usage{
rank_sale(sale, rank_var = "duration", grp_var = c("cust_id", "year"))
}
\arguments{
\item{sale}{data frame: Input sales data}

\item{rank_var}{character: name of variable(s) to use for ranking}

\item{grp_var}{character: name of variable(s) used for grouping}
}
\description{
The filter is based on one or more variables (rank_var); the row with the maximum 
value(s) of rank_var is selected. The default (and intended purpose) is to 
pick the maximum "duration" value per customer-year. This is a preliminary step
to \code{\link{make_lic_history}} which ensures that multi-year & lifetime sales
are always accounted for.
}
\examples{
library(dplyr)
data(lic, sale)
sale_unranked <- left_join(sale, lic)
sale_ranked <- rank_sale(sale_unranked) \%>\%
    join_first_month(sale_unranked)
    
# check sale ranking - highest duration will always be picked
left_join(
    count(sale_ranked, duration), 
    distinct(sale_unranked, cust_id, year, duration) \%>\% count(duration), 
    by = "duration",
    suffix = c(".ranked", ".unranked")
)
}
\seealso{
Other license history functions: \code{\link{carry_duration}},
  \code{\link{carry_variables}},
  \code{\link{check_history_samp}},
  \code{\link{check_identify_R3}},
  \code{\link{check_identify_lapse}},
  \code{\link{identify_R3}}, \code{\link{identify_lapse}},
  \code{\link{join_first_month}},
  \code{\link{make_lic_history}}
}
\concept{license history functions}
