% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_agecat}
\alias{recode_agecat}
\title{Create a standardized age category variable}
\usage{
recode_agecat(dat, age_labs = c("0-17", "18-24", "25-34", "35-44", "45-54",
  "55-64", "65+"), age_breaks = c(-Inf, 17, 24, 34, 44, 54, 64, Inf),
  max_age = 110)
}
\arguments{
\item{dat}{data frame: Input table}

\item{age_labs}{character: labels to use for age category}

\item{age_breaks}{numeric: breaks for age category passed to 
\code{\link[base]{cut}}}

\item{max_age}{numeric: maximum allowed age. Anything above will be set to missing.}
}
\description{
Uses 'birth_year' (from cust) and 'year' (from sale) to identify
age for each license year
}
\examples{
data(cust, sale, package = "salic")
sale2 <- left_join(sale, cust) \%>\% 
    mutate(birth_year = lubridate::year(dob)) \%>\% 
    recode_agecat()
}
\seealso{
Other functions for recoding license data: \code{\link{recode_month}},
  \code{\link{recode_state}}
}
