% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_month}
\alias{recode_month}
\title{Create a standardized month variable for sales table}
\usage{
recode_month(sale, month_range = 0:12)
}
\arguments{
\item{sale}{data frame: Input sales table}

\item{month_range}{numeric: A vector of months allowed in the output. Defaults
to 0:12 since this is a common range of sales. Any months outside the range will be
set to either the lowest values (for those less than the range) or the
highest value (for those over). This prevents unusual sale months from appearing
in results.}
}
\value{
Returns a sales table with a standardized 'month' variable and
prints a validation output.
}
\description{
The output month variable capture transaction year and month in one metric.
The month number is relative to the license year: 1 = Jan, 2 = Feb, etc, for
current/future years and 0 = Dec, -1 = Nov, ..., for previous years.
}
\details{
Assumes two variables are present in the sales table:
dot (transaction date in 'yyyy-mm-dd'), year (numeric license year)
}
\seealso{
Other functions for recoding license data: \code{\link{recode_agecat}},
  \code{\link{recode_state}}
}
\concept{functions for recoding license data}
