% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{recode_state}
\alias{recode_state}
\title{Recode state abbreviations}
\usage{
recode_state(dat, state_table, oldvar = "state", newvar = "state_new")
}
\arguments{
\item{dat}{data frame: input data with a state variable}

\item{state_table}{data frame: input data holding valid abbreviations}

\item{oldvar}{character: name of state variable}

\item{newvar}{character: name of new state variable}
}
\description{
This is a convenience function to help standardize odd-looking state abbreviations
based on a reference list of allowed abbreviations.
}
\examples{
data(state_abbreviations)
dat <- data.frame(state = c("M0", "mo", "Mo", "MR", "YY"))
recode_state(dat, state_abbreviations)
}
\seealso{
Other functions for standardizing state license data: \code{\link{recode_month}}
}
\concept{functions for standardizing state license data}
