% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{replace_strings}
\alias{replace_strings}
\title{Search and replace string across files with R}
\usage{
replace_strings(path = ".", find_string, replacement_string,
  pattern = ".R", showmessage = TRUE)
}
\arguments{
\item{path}{a character vector of full path names; the default
    corresponds to the working directory, \code{\link{getwd}()}.  Tilde
    expansion (see \code{\link{path.expand}}) is performed.  Missing
    values will be ignored.}

\item{find_string}{character: String (to find) that will be replaced}

\item{replacement_string}{character: New string to use}

\item{pattern}{an optional \link{regular expression}.  Only file names
    which match the regular expression will be returned.}

\item{showmessage}{logical: If TRUE, prints a message about replacement}
}
\description{
This is a helper function for preparing templates in salic. It uses 
\code{\link[base]{gsub}} (https://gist.github.com/mages/1544009)
}
\examples{
salic::replace_strings()
}
\keyword{internal}
