% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{sale}
\alias{sale}
\title{Sample Data: Hunting/Fishing Sales for 1,000 Customers}
\format{A data frame with 4,651 rows and 5 variables
\describe{
\item{cust_id}{Customer ID (integer) (foreign key)}
\item{lic_id}{License ID (integer) (foreign key)}
\item{year}{License Year (integer) (yyyy)}
\item{dot}{Transaction Purchase Date (character) (yyyy-mm-dd)}
\item{res}{Residency (integer) (1=Res, 0=Nonres)}
}}
\description{
Sample Data: Hunting/Fishing Sales for 1,000 Customers
}
\seealso{
Other Sample Data: \code{\link{cust}}, \code{\link{lic}},
  \code{\link{state_abbreviations}}
}
\concept{Sample Data}
\keyword{datasets}
