% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{samp_dups}
\alias{samp_dups}
\title{Sample duplicates in a table}
\usage{
samp_dups(dup_all, id, size = 10)
}
\arguments{
\item{dup_all}{data frame: duplicate values joined to customer table,
output of get_dups}

\item{id}{character: Variable(s) representing identifier to check}
}
\description{
Intended to be run following get_dups
}
\examples{
library(dplyr)
cust <- data.frame(id = 100001:100003, nm = c("dan k","dan k","dan r"))
x <- select(cust, id, nm) \%>\% distinct()
select(x, nm) \%>\% check_dups()
get_dups(x, cust, "nm") \%>\%
    samp_dups("nm")
}
\seealso{
Other functions for checking license data: \code{\link{check_cust}},
  \code{\link{check_dups}}, \code{\link{get_dups}},
  \code{\link{show_dups}}
}

