% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{show_dups}
\alias{show_dups}
\title{Collect duplicates in a table}
\usage{
show_dups(x)
}
\arguments{
\item{x}{data frame: Table holding records}
}
\description{
Collect duplicates in a table
}
\examples{
# this example requires a dataset with names populated
library(dplyr)
x <- select(cust, cust_id, dob, last, first) \%>\% distinct()
dup <- select(x, dob, last, first) \%>\% show_dups()
}
\seealso{
Other functions for checking license data: \code{\link{check_cust}},
  \code{\link{check_dups}}, \code{\link{get_dups}},
  \code{\link{samp_dups}}
}

