% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\docType{data}
\name{state_abbreviations}
\alias{state_abbreviations}
\title{Data: State/Province/Territory Abbreviations for US & Canada}
\format{A data frame with 72 rows and 3 variables
\describe{
\item{name}{State/Province/Territory name}
\item{state}{Abbreviation}
\item{country}{Country Abbreviation where CN = Canada}
}}
\description{
A data frame with 72 abbreviations, used in the \code{\link{recode_state}} function
}
\seealso{
Other functions for standardizing state license data: \code{\link{recode_month}},
  \code{\link{recode_state}}
}
\concept{functions for standardizing state license data}
\keyword{datasets}
