% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{summary_initial}
\alias{summary_initial}
\title{Summarize sales, customers, and churn by year}
\usage{
summary_initial(sale)
}
\arguments{
\item{sale}{data frame: Table holding sales by year with a minimum of
2 variables (cust_id, year)}
}
\description{
This is a wrapper function that combines \code{\link{summary_sale}} and 
\code{\link{summary_churn}}
}
\examples{
library(tidyverse)
library(salic)
data(sale, lic, package = "salic")

summary_initial(sale)
}
\seealso{
Other functions for validating license data: \code{\link{check_dups}},
  \code{\link{count_lines_textfile}},
  \code{\link{summary_churn}}, \code{\link{summary_sale}}
}
\concept{functions for validating license data}
