% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{summary_sale}
\alias{summary_sale}
\title{Summarize sales and customers by year}
\usage{
summary_sale(x, include_revenue = FALSE, rnd = 1, out = NULL,
  title = "Summarize Sales & Customers", note = NULL)
}
\arguments{
\item{x}{data frame: Table holding sales by year with a minimum of
2 variables (cust_id, year)}

\item{include_revenue}{logical: If TRUE, revenue by year will also be shown}

\item{rnd}{numeric: number of decimals to round pct change results}

\item{out}{character: file path to optional output csv}

\item{title}{character: title for output table}

\item{note}{character: note for output table}
}
\description{
Summarize sales and customers by year
}
\examples{
summary_sale(sale)
summary_sale(sale, out = "summary_sale.csv")

library(dplyr)
sale <- left_join(sale, lic)
filter(sale, priv == "hunt") \%>\%
    summary_sale(title = "Hunting Sales/Cust", note = "An additional note")

# example with made-up license revenue
sale <- mutate(sale, revenue = 30)
summary_sale(sale, rnd = 2, include_revenue = T)
}
\seealso{
Other functions for validating license data: \code{\link{calc_churn}},
  \code{\link{cust_summary_demo}},
  \code{\link{lic_summary_revenue}},
  \code{\link{lic_summary}}, \code{\link{pct_round}},
  \code{\link{summary_churn}}
}

