% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{track}
\alias{track}
\title{Sample Data: GA Hunting purchase History}
\format{A list of data frames, each with 1000 rows and 3 variables
\describe{
\item{cust_id}{Customer ID}
\item{left}{Time remaining on previous license (==0: need to renew,
>0: don't need to renew, ==NA: haven't bought before, <0 didn't buy last year)}
\item{duration}{Bought a license this year? (>=1: yes, ==0: no)}
}}
\description{
A list of data frames corresponding to the 1000 customer sample
}
\seealso{
Other Sample Data: \code{\link{cust}}, \code{\link{lic}},
  \code{\link{sale}}, \code{\link{state_abbreviations}}
}
\keyword{datasets}

