% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history-deprecated.R
\name{update_track}
\alias{update_track}
\title{Update license history data for a single year}
\usage{
update_track(x, sale_ranked, yr, carry_vars = NULL)
}
\arguments{
\item{sale_ranked}{data frame: sale table showing purchases for all years}

\item{yr}{numeric: current year, used to subset sale data for updating tracking}

\item{carry_vars}{character: variables to carry over from previous year
(for multi-year and lifetime licenses)}

\item{track}{data frame: tracking table with a variable ('left') which is
updated based on purchases in the current year}
}
\description{
This is called as a part of \code{\link{make_track}}
}
\examples{
update_track()
}
\seealso{
Other deprecated license history functions: \code{\link{check_R3}},
  \code{\link{check_lapse}}, \code{\link{check_make_priv}},
  \code{\link{check_make_track}},
  \code{\link{check_track}}, \code{\link{get_R3}},
  \code{\link{get_lapse}}, \code{\link{make_priv_final}},
  \code{\link{make_priv}}, \code{\link{make_track}}
}

