% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-check.R
\name{variable_allowed_values}
\alias{variable_allowed_values}
\title{Internal Function: Check allowed values for a single variable}
\usage{
variable_allowed_values(x, x_name, allowed_values)
}
\arguments{
\item{x}{vector: variable to check}

\item{x_name}{character: name of variable to print in warning}

\item{allowed_values}{vector: values the variable is allowed to take}
}
\description{
This is intended to be called from \code{\link{data_allowed_values}}. 
Prints a warning message on a failed check.
}
\examples{
data(lic)
variable_allowed_values(lic$type, "lic-type", c("hunt", "fish", "combo"))

x <- c(NA, "hi", 1, lic$type)
variable_allowed_values(x, "lic-type", c("hunt", "fish", "combo"))
}
\seealso{
Other functions to check data format: \code{\link{data_check_table}},
  \code{\link{data_foreign_key}},
  \code{\link{data_primary_key}}
}
\concept{functions to check data format}
\keyword{internal}
