% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{warn}
\alias{warn}
\title{Helper function: print a warning to flag records}
\usage{
warn(flagged, msg)
}
\arguments{
\item{flagged}{data frame: table containing flagged rows}

\item{msg}{character: message to be printed in warming}
}
\description{
The warning is only printed if the input data frame has more than one row 
(i.e., only flagged rows should be passed to warn())
}
\examples{
x <- data.frame(x = 1:2, val = c("careful", "a bit high"))
warn(x, "I'm a warning")
}
\concept{dashboard functions}
\keyword{internal}
