% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{carry_variables}
\alias{carry_variables}
\title{Internal Function: Carry specified variables forward}
\usage{
carry_variables(sale_split, yrs, carry_vars)
}
\arguments{
\item{sale_split}{list: ranked sale table split by year}

\item{yrs}{numeric: Years in sales data (column 'year') from which
to create license history}

\item{carry_vars}{character: additional variables to carry over from previous year
(for multi-year and lifetime licenses).}
}
\description{
This function is intended to be called from \code{\link{make_lic_history}}.
}
\examples{
library(dplyr)
data(sale, lic)

sale_unranked <- left_join(lic, sale)
sale_ranked <- rank_sale(sale_unranked)

split(sale_ranked, sale_ranked$year) \%>\%     
    carry_duration(2008:2019) \%>\%
    carry_variables(2008:2019, c("month", "res"))
}
\seealso{
Other license history functions: \code{\link{carry_duration}},
  \code{\link{check_history_samp}},
  \code{\link{check_identify_R3}},
  \code{\link{check_identify_lapse}},
  \code{\link{identify_R3}}, \code{\link{identify_lapse}},
  \code{\link{join_first_month}},
  \code{\link{make_lic_history}}, \code{\link{prep_yrs}},
  \code{\link{rank_sale}}
}
\concept{license history functions}
\keyword{internal}
