% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_dups}
\alias{check_dups}
\title{Check for duplicates in a table}
\usage{
check_dups(x)
}
\arguments{
\item{x}{data frame: Table holding records}
}
\description{
Check for duplicates in a table
}
\examples{
library(dplyr)
cust <- data.frame(id = 1:3, nm = c("dan k","dan k","dan r"))
x <- select(cust, id, nm) \%>\% distinct()
select(x, nm) \%>\% check_dups()
}
\seealso{
Other functions for validating license data: \code{\link{count_lines_textfile}},
  \code{\link{summary_churn}},
  \code{\link{summary_initial}}, \code{\link{summary_sale}}
}
