% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{check_history_samp}
\alias{check_history_samp}
\title{Sample the output of \code{\link{make_lic_history}}}
\usage{
check_history_samp(lic_history, n_samp = 3, buy_min = 3, buy_max = 8)
}
\arguments{
\item{lic_history}{license history data frame produced with \code{\link{make_lic_history}}}

\item{n_samp}{numeric: number of customers to view}

\item{buy_min}{numeric: minimum number of license purchases for customers to include}

\item{buy_max}{numeric: maximum number of license purchases for customers to include}
}
\description{
View a sample of customers from license history table to check the 
year over year dynamics (outputs a list split by customer ID).
}
\examples{
data(history)
check_history_samp(history, n_samp = 5)
}
\seealso{
Other license history functions: \code{\link{carry_duration}},
  \code{\link{carry_variables}},
  \code{\link{check_identify_R3}},
  \code{\link{check_identify_lapse}},
  \code{\link{identify_R3}}, \code{\link{identify_lapse}},
  \code{\link{join_first_month}},
  \code{\link{make_lic_history}}, \code{\link{rank_sale}}
}
\concept{license history functions}
