% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{check_identify_R3}
\alias{check_identify_R3}
\title{Check the output of \code{\link{identify_R3}}}
\usage{
check_identify_R3(lic_history, yrs)
}
\arguments{
\item{lic_history}{license history data frame produced with \code{\link{make_lic_history}}}

\item{yrs}{numeric: Years in sales data (column 'year') from which
to create license history}
}
\description{
A test to insure R3 was correctly coded
}
\examples{
library(dplyr)
data(history)

select(history, -R3) \%>\%
    identify_R3(2008:2019, show_check_vars = TRUE) \%>\%
    check_identify_R3(2008:2019) \%>\%
    data.frame()
}
\seealso{
Other license history functions: \code{\link{check_history_samp}},
  \code{\link{check_identify_lapse}},
  \code{\link{identify_R3}}, \code{\link{identify_lapse}},
  \code{\link{join_first_month}},
  \code{\link{make_lic_history}}, \code{\link{rank_sale}}
}
\concept{license history functions}
