% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-check.R
\name{check_lapse}
\alias{check_lapse}
\title{Check privilege table (lapse)}
\usage{
check_lapse(x)
}
\arguments{
\item{x}{list: license privilege tables returned by \code{\link{make_priv}}}
}
\description{
Check privilege table (lapse)
}
\examples{
check_lapse()
}
\seealso{
Other functions for checking license history summarization: \code{\link{check_R3}},
  \code{\link{check_make_priv_final}},
  \code{\link{check_make_priv}},
  \code{\link{check_make_track}}, \code{\link{check_track}}
}

