% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history-deprecated.R
\name{check_make_priv}
\alias{check_make_priv}
\title{Check Privilege Table Creation}
\usage{
check_make_priv(priv)
}
\arguments{
\item{priv}{list: license privilege tables returned by \code{\link{make_priv}}}
}
\description{
Check Privilege Table Creation
}
\examples{
check_make_priv()
}
\seealso{
Other deprecated license history functions: \code{\link{check_R3}},
  \code{\link{check_lapse}},
  \code{\link{check_make_track}},
  \code{\link{check_track}}, \code{\link{get_R3}},
  \code{\link{get_lapse}}, \code{\link{make_priv_final}},
  \code{\link{make_priv}}, \code{\link{make_track}},
  \code{\link{update_track}}
}

