% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{check_make_priv_final}
\alias{check_make_priv_final}
\title{Check Final Privilege Table Creation}
\usage{
check_make_priv_final(priv, rnd = 1)
}
\arguments{
\item{priv}{data frame: license privilege table}

\item{rnd}{numeric: number of decimals to round pct change results}
}
\description{
Produces a summary of the priv table with (1) total counts, (2) lapse, and (3) R3
}
\examples{
### Run ranking with sample data
library(tidyverse)
library(salic)
data(lic, sale, package = "salic")

sale_unranked <- select(lic, lic_id, duration) \%>\%
    right_join(sale) \%>\%
    select(cust_id, year, duration)
sale_ranked <- rank_sale(sale_unranked)

### Make license history
yrs <- 2004:2013
lic_history <- make_lic_history(sale_ranked, yrs)

### Identify R3 & lapse
lic_history <- identify_R3(lic_history, yrs)
lic_history <- identify_lapse(lic_history, yrs)

# make priv table & check
priv <- lic_history \%>\%
    filter(has_priv) \%>\%
    select(cust_id, year, R3, lapse)
check_make_priv_final(priv)
}
\seealso{
Other license history functions: \code{\link{check_history_samp}},
  \code{\link{check_identify_R3}},
  \code{\link{check_identify_lapse}},
  \code{\link{check_rank_sale}},
  \code{\link{ggplot_lic_history}},
  \code{\link{identify_R3}}, \code{\link{identify_lapse}},
  \code{\link{make_lic_history}}, \code{\link{rank_sale}}
}

