% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history-deprecated.R
\name{check_R3}
\alias{check_R3}
\title{Check privilege table (R3)}
\usage{
check_R3(x)
}
\arguments{
\item{x}{list: license privilege tables returned by \code{\link{make_priv}}}
}
\description{
Check privilege table (R3)
}
\examples{
check_R3()
}
\seealso{
Other deprecated license history functions: \code{\link{check_lapse}},
  \code{\link{check_make_priv}},
  \code{\link{check_make_track}},
  \code{\link{check_track}}, \code{\link{get_R3}},
  \code{\link{get_lapse}}, \code{\link{make_priv_final}},
  \code{\link{make_priv}}, \code{\link{make_track}},
  \code{\link{update_track}}
}

