% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-check.R
\name{check_rank_sale}
\alias{check_rank_sale}
\title{Check the output of \code{\link{rank_sale}}}
\usage{
check_rank_sale(sale_ranked, sale_unranked, check_duration = TRUE)
}
\arguments{
\item{sale_ranked}{data frame: Data produced after running
\code{\link{rank_sale}}}

\item{sale_unranked}{data frame: Data used as input to \code{\link{rank_sale}}}

\item{check_duration}{logical: If TRUE, runs a check to see if any
customers have mutliple license sales (with different durations) in the same
year (and insures these are ranked correctly if any are found). If this isn't
a concern, set to FALSE to avoid this relatively time-consuming test.}
}
\description{
A test to insure that sale ranking was performed correctly
}
\examples{
### Test ranking on sample data
library(dplyr)
library(tidyr)

# Join duration variable to sale data
sale_unranked <- select(lic, lic_id, duration) \%>\%
    right_join(sale) \%>\%
    select(cust_id, year, duration)

# Test Ranking
sale_ranked <- rank_sale(sale_unranked)
check_rank_sale(sale_ranked, sale_unranked)
}
\seealso{
Other functions for for checking license history summarization: \code{\link{catf}},
  \code{\link{printf}}
}

