% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-check.R
\name{check_track}
\alias{check_track}
\title{Check license history (stored in a list)}
\usage{
check_track(x)
}
\arguments{
\item{x}{list: license history tables returned by \code{\link{make_track}}}
}
\description{
Check license history (stored in a list)
}
\examples{
check_track()
}
\seealso{
Other functions for checking license history summarization: \code{\link{check_R3}},
  \code{\link{check_lapse}},
  \code{\link{check_make_priv_final}},
  \code{\link{check_make_priv}},
  \code{\link{check_make_track}}
}

