% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{count_lines_textfile}
\alias{count_lines_textfile}
\title{Count rows in a text file}
\usage{
count_lines_textfile(file_path)
}
\arguments{
\item{file_path}{character: Path to file}
}
\description{
This is a quick way of counting lines in a text file, to insure the final # of rows in 
the loaded dataset is correct: 
 https://stackoverflow.com/questions/23456170/get-the-number-of-lines-in-a-text-file-using-r
}
\examples{
count_lines_textfile
}
\seealso{
Other functions for validating license data: \code{\link{check_dups}},
  \code{\link{summary_churn}},
  \code{\link{summary_initial}}, \code{\link{summary_sale}}
}
\concept{functions for validating license data}
