% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{cust}
\alias{cust}
\title{Sample Data: 30,000 Customers}
\format{A data frame with 3 variables
\describe{
\item{cust_id}{Customer ID (numeric) (primary key) - Note that IDs can be stored 
as character, but this is less efficient.}
\item{sex}{Gender (numeric) (1=Male, 2=Female, NA=Unknown)}
\item{birth_year}{Year of Birth (numeric) (yyyy, NA=Unknown)}
}}
\description{
Sample Data: 30,000 Customers
}
\seealso{
Other Sample Data: \code{\link{dashboard}},
  \code{\link{data_salic}}, \code{\link{history}},
  \code{\link{lic}}, \code{\link{metrics}},
  \code{\link{sale}}, \code{\link{state_abbreviations}}
}
\concept{Sample Data}
\keyword{datasets}
