% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{cust}
\alias{cust}
\title{Sample Data: 25,000 Customers}
\format{A data frame with 3 variables
\describe{
\item{cust_id}{Customer ID (integer) (primary key)}
\item{sex}{Gender (integer) (1=Male, 2=Female)}
\item{birth_year}{Year of Birth (integer) (yyyy)}
}}
\description{
Sample Data: 25,000 Customers
}
\seealso{
Other Sample Data: \code{\link{dashboard}},
  \code{\link{history}}, \code{\link{lic}},
  \code{\link{metrics}}, \code{\link{sale}},
  \code{\link{state_abbreviations}}
}
\concept{Sample Data}
\keyword{datasets}
