% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{cust}
\alias{cust}
\title{Sample Data: GA Hunting Customers}
\format{A data frame with 1000 rows and 4 variables
\describe{
\item{cust_id}{Customer ID}
\item{dob}{Date of Birth}
\item{state}{State of Residence}
\item{sex}{Gender}
}}
\description{
A data frame with a random sample of 1000 GA hunting customers
}
\seealso{
Other Sample Data: \code{\link{lic}}, \code{\link{sale}},
  \code{\link{state_abbreviations}}
}
\keyword{datasets}
