% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-check.R
\name{data_check_table}
\alias{data_check_table}
\alias{data_check_cust}
\alias{data_check_lic}
\alias{data_check_sale}
\title{Check formatting rules for standardized data tables}
\usage{
data_check_table(df, df_name, primary_key, required_vars, allowed_values)

data_check_cust(df, df_name = "cust", primary_key = "cust_id",
  required_vars = c("cust_id", "sex", "birth_year"),
  allowed_values = list(sex = c(1, 2, NA), birth_year =
  c(1900:substr(Sys.Date(), 1, 4), NA)))

data_check_lic(df, df_name = "lic", primary_key = "lic_id",
  required_vars = c("lic_id", "type", "duration"),
  allowed_values = list(type = c("fish", "hunt", "combo"), duration =
  1:99))

data_check_sale(df, df_name = "sale", primary_key = NULL,
  required_vars = c("cust_id", "lic_id", "year", "month", "res"),
  allowed_values = list(year = c(2000:substr(Sys.Date(), 1, 4)), month =
  1:12, res = c(1, 0, NA)))
}
\arguments{
\item{df}{data frame: table to check}

\item{df_name}{character: name of relevant data table ("cust", "lic", or "sale")}

\item{primary_key}{character: name of variable that acts as primary key,
which should be unique and non-missing. NULL indicates no primary key in table.}

\item{required_vars}{character: variables that should be included}

\item{allowed_values}{list: named list with allowed values for specific variables}
}
\description{
Prints a warning if any of the specied formatting rules don't pass (silent otherwise). 
Table-specific versions are convenience functions that call data_check_table() 
with appropriate defaults.
}
\details{
Developer note: data_check_table() is itself a wrapper 
for several internal functions (see \code{\link{data_internal}}).
}
\examples{
library(dplyr)

# produce various format warnings
data(cust)
bind_rows(cust, cust) \%>\% data_check_cust()

cust$birth_year[1] <- 2100
data_check_cust(cust)

data(lic)
select(lic, -duration) \%>\% data_check_lic()
mutate(lic, duration = 0) \%>\% data_check_lic()

data(sale)
sale$year[1] <- NA
data_check_sale(sale)
}
\seealso{
Other functions to check data format: \code{\link{data_check}},
  \code{\link{data_foreign_key}},
  \code{\link{data_internal}},
  \code{\link{variable_allowed_values}}
}
\concept{functions to check data format}
\keyword{internal}
