% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-check.R
\name{data_primary_key}
\alias{data_primary_key}
\alias{data_required_vars}
\title{Internal Functions: Individual Data Checks}
\usage{
data_primary_key(df, df_name, primary_key)

data_required_vars(df, df_name, required_vars)
}
\arguments{
\item{df}{data frame: table to check}

\item{df_name}{character: name of relevant data table (e.g., "lic", "sale", etc.)}

\item{primary_key}{character: name of variable that acts as primary key (if applicable)}

\item{required_vars}{character: variables that should be included}
}
\description{
These functions are intended to be called from \code{\link{data_check_table}}. 
Prints a warning message on a failed check.
}
\examples{
library(dplyr)
data(lic)

bind_rows(lic, lic) \%>\% 
    data_primary_key("lic", "lic_id")
    
select(lic, -duration) \%>\%
    data_required_vars("lic", c("lic_id", "type", "duration"))
}
\seealso{
Other functions to check data format: \code{\link{data_check_table}},
  \code{\link{data_foreign_key}}
}
\concept{functions to check data format}
\keyword{internal}
