% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\name{data_salic}
\alias{data_salic}
\title{Salic Sample Data Overview}
\description{
Salic provides sample data for each stage of dashboard production:
\itemize{
  \item Standardized License Data
    \itemize{
      \item Customers: \code{\link{cust}}
      \item License Types: \code{\link{lic}}
      \item Sales: \code{\link{sale}}
    }
  \item License History Data: \code{\link{history}}
  \item Dashboard Metrics
    \itemize{
      \item Calculated dashboard metrics (stored in list): \code{\link{metrics}}
      \item Metrics formatted for Dashboard Input (single table): \code{\link{dashboard}}
    }
}
}
\seealso{
Salic Function Reference: \code{\link{salic}}

Other Sample Data: \code{\link{cust}},
  \code{\link{dashboard}}, \code{\link{history}},
  \code{\link{lic}}, \code{\link{metrics}},
  \code{\link{sale}}
}
\concept{Sample Data}
