% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{est_part}
\alias{est_part}
\title{Estimate participants by year from license history}
\usage{
est_part(history, segment = "tot", test_threshold = 30,
  include_test_value = FALSE)
}
\arguments{
\item{history}{data frame: input license history table}

\item{segment}{character: defaults to "tot", which indicates no segmentation.
Alternatively specifiy other license history variables ("res", "sex", etc.)}

\item{test_threshold}{numeric: threshold in whole number percentage points 
for pct. change per year. A warning will be printed if the absolute value
of the change for any year exceeds the threshold.}

\item{include_test_value}{logical: If TRUE, the output table will include
a variable holding the test value (pct. change per year) for each row.}
}
\description{
This function requires a correctly formated history table (see \code{\link{history}}).
It produces a simple count of records per year, optionally by segment, 
& runs a validation test: pct. change per year. A large pct. change in a year
can often indicate a problem.
}
\examples{
library(dplyr)
data(history)
history <- label_categories(history) \%>\%
    recode_agecat() \%>\%
    filter(!agecat \%in\% c("0-17", "65+"))

# a flag will be raised since 2019 is a partial year
est_part(history, include_test_value = TRUE)

# fix by dropping partial year
history <- filter(history, year != 2019)
est_part(history)

# by age category
est_part(history, "agecat")
est_part(history, "agecat", test_threshold = 35)

# apply over multiple segments
segs <- c("tot", "res", "sex", "agecat")
sapply(segs, function(x) est_part(history, x), simplify = FALSE)
}
\seealso{
Other dashboard functions: \code{\link{warn}}
}
\concept{dashboard functions}
