% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-factor.R
\name{factor_var}
\alias{factor_var}
\alias{factor_age}
\alias{factor_sex}
\alias{factor_res}
\alias{factor_R3}
\title{Convert numeric variable to factor}
\usage{
factor_var(x, levels, labels, ...)

factor_age(x, levels = 1:7, labels = c("0-17", "18-24", "25-34",
  "35-44", "45-54", "55-64", "65+"), ...)

factor_sex(x, levels = 1:2, labels = c("Male", "Female"), ...)

factor_res(x, levels = c(1, 0), labels = c("Resident", "Nonresident"),
  ...)

factor_R3(x, levels = 1:4, labels = c("Carry", "Retain", "Reactivate",
  "Recruit"), ...)
}
\arguments{
\item{x}{numeric: Input numeric vector}

\item{levels}{numeric: Levels for input numeric vector}

\item{labels}{labels: Labels to use for output factor vector}

\item{...}{Other arguments passed to \code{\link[base]{factor}}}
}
\description{
These are convenience functions for use with license data. They are
wrappers for \code{\link[base]{factor}} with specified values.
}
\examples{
library(dplyr)
data(cust, sale, package = "salic")
sale2 <- left_join(sale, cust) \%>\% 
    recode_agecat()
sale2 \%>\%
    select(-agecat) \%>\%
    mutate(agecat = factor_age(age))
}
\seealso{
Other functions for labelling numeric variables: \code{\link{df_factor_var}}
}
\concept{functions for labelling numeric variables}
