% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{format_tableau}
\alias{format_tableau}
\title{Format estimated metrics for input to Tableau Dashboard}
\usage{
format_tableau(df, timeframe, group, rename_input = c("category", "year",
  "value"))
}
\arguments{
\item{df}{data frame: Input table (3 variables) with estimated metrics}

\item{timeframe}{character: value to store in the 'timeframe' variable of 
the output (e.g, 'full-year', 'mid-year')}

\item{group}{character: value to sore in the 'group' variable of the 
output (e.g., 'all_sports', 'fish', 'hunt')}

\item{rename_input}{character: generic names for input variables as they
will appear in the output}
}
\description{
This function expects a data frame produced by an salic estimation function (
\code{\link{est_part}}, \code{\link{est_recruit}}, \code{\link{est_churn}}).
It returns a data frame with additional formatting that allows stacking all results
into a single table.
}
\examples{
library(dplyr)
data(metrics)

# format a table
metrics$participants$res
x <- format_tableau(metrics$participants$res, "full-year", "all_sports")
x

# combine formatted tables
y <- format_tableau(metrics$participants$tot, "full-year", "all_sports")
bind_rows(y, x)

# apply formatting across all segments
x <- lapply(metrics$participants, function(x) format_tableau(x, "full-year", "sports"))
bind_rows(x)
   
# apply across all metrics & segments
bind_rows(
    lapply(metrics$participants, function(x) format_tableau(x, "full-year", "sports")),
    lapply(metrics$recruits, function(x) format_tableau(x, "full-year", "sports")),
    lapply(metrics$churn, function(x) format_tableau(x, "full-year", "sports"))
)
}
\seealso{
Other dashboard functions: \code{\link{check_threshold}},
  \code{\link{est_churn}}, \code{\link{est_part}},
  \code{\link{scaleup_part}}
}
\concept{dashboard functions}
