% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{get_R3}
\alias{get_R3}
\title{Identify R3 group each year}
\usage{
get_R3(track, i)
}
\arguments{
\item{track}{list: Set of license history tables created using
\code{\link{make_track}}}

\item{i}{integer/character: position of data frame in tracking list to
calculate metric}
}
\description{
This is called as part of \code{\link{make_priv}} and codes R3: 1 = carried,
2 = retained, 3 = reactivated, 4 = recruited, where 1 (retained) consists
of carried + renewed.
}
\details{
R3 Conditionals:
\itemize{
\item \emph{1 (carry)}: License doesn't expire this year
(l!is.na(left) & left > 0)
\item \emph{2 (renew)}: Renewed a license this year
(duration >=1 & !is.na(left) & left == 0)
\item \emph{3 (reactivate)}: Bought a license, no privilege last year but had
a privilege within the past 5 years
(duration >=1 & !is.na(left) & left < 0 & left > -5)
\item \emph{4 (recruit)}: Bought a license, but no privilege in past 5 years
(duration >=1 & (is.na(left) | left <= -5))
}
}
\examples{
# Calculate R3 in 2012
library(dplyr)
x <- get_R3(track, "2012")

# Show counts of customers who bought licenses in 2012
# Recruited (R3=4) bought never (NA) or 6 or more years previously
# Reactivated (R3=3) bought between 2 and 5 years before
# Renewed (R3=2) bought last year
# Carried (R3=1) bought a multi-year or lifetime license
count(x, duration, left, R3) \%>\%
    filter(!is.na(R3)) \%>\%
    tidyr::spread(duration, n, fill = "")
}
\seealso{
Other functions for summarizing license history: \code{\link{get_lapse}},
  \code{\link{make_priv_final}}, \code{\link{make_priv}},
  \code{\link{make_track}}, \code{\link{rank_sale}},
  \code{\link{update_track}}
}

