% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{forward_duration}
\alias{forward_duration}
\alias{forward_vars}
\alias{make_lapse}
\alias{make_R3}
\alias{history_internal}
\title{Internal Functions: Making license history}
\usage{
forward_duration(df, df_last, current_year, carry_vars)

forward_vars(df, carry_vars = NULL)

make_lapse(df_last, df, yrs_lapse)

make_R3(df, include_R3, yrs)
}
\arguments{
\item{df}{data frame: current year table}

\item{df_last}{data frame: previous year table}

\item{carry_vars}{character: additional variables to carry over from previous year
(for multi-year and lifetime licenses).}

\item{yrs_lapse}{numeric: years to include in lapse calculation (defaults to yrs). 
If NULL, lapse will not be calculated (useful for mid-year results)}

\item{include_R3}{logical: If TRUE, also calculate R3}

\item{yrs}{numeric: Years in sales data (column 'year') from which
to create license history}
}
\description{
These functions are intended to be called from \code{\link{make_history}}.
}
\examples{
# example
}
\seealso{
Other license history functions: \code{\link{join_first_month}},
  \code{\link{make_history}}, \code{\link{prep_yrs}},
  \code{\link{rank_sale}}
}
\concept{license history functions}
\keyword{internal}
