% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{history_internal}
\alias{history_internal}
\alias{forward_duration}
\alias{forward_vars}
\alias{make_lapse}
\alias{make_R3}
\title{Internal Functions: Making license history}
\usage{
forward_duration(dt, current_year)

forward_vars(dt, carry_vars)

make_lapse(dt_last, dt)

make_R3(dt, yrs)
}
\arguments{
\item{dt}{data.table: current year table}

\item{current_year}{numeric: year of current year table}

\item{carry_vars}{character: additional variables to carry over from previous year
(for multi-year and lifetime licenses).}

\item{dt_last}{data.table: previous year table}

\item{yrs}{numeric: Years in sales data (column 'year') from which
to create license history}
}
\description{
These functions are only to be called from \code{\link{make_history}}. 
They work by side-effects; modifying the input data.table by reference.
}
\examples{
# examples
}
\seealso{
Other license history functions: \code{\link{join_first_month}},
  \code{\link{make_history}}, \code{\link{prep_yrs}},
  \code{\link{rank_sale}}
}
\concept{license history functions}
\keyword{internal}
