% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{identify_lapse}
\alias{identify_lapse}
\title{Identify lapse group each year}
\usage{
identify_lapse(lic_history, yrs)
}
\description{
This is intended to be called following \code{\link{make_lic_history}} 
and codes lapse: 0 = renew next year, 1 = lapse next year.
}
\examples{
### Run ranking with sample data
library(tidyverse)
library(salic)
data(lic, sale, package = "salic")

sale_unranked <- select(lic, lic_id, duration) \%>\%
    right_join(sale) \%>\%
    select(cust_id, year, duration)
sale_ranked <- rank_sale(sale_unranked)

### Make license history
yrs <- 2004:2013
lic_history <- make_lic_history(sale_ranked, yrs)

### Identify lapse
lic_history <- identify_lapse(lic_history, yrs)

# check
ggplot_lic_history(lic_history, yrs, "lapse")
check_identify_lapse(lic_history)
check_history_samp(lic_history)
}
\seealso{
Other license history functions: \code{\link{check_history_samp}},
  \code{\link{check_history_summary}},
  \code{\link{check_identify_R3}},
  \code{\link{check_identify_lapse}},
  \code{\link{check_make_priv_final}},
  \code{\link{check_rank_sale}},
  \code{\link{ggplot_lic_history}},
  \code{\link{identify_R3}},
  \code{\link{make_lic_history}}, \code{\link{rank_sale}}
}

