% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{identify_lapse}
\alias{identify_lapse}
\title{Identify lapse group each year}
\usage{
identify_lapse(lic_history, yrs = sort(unique(lic_history$year)),
  show_summary = FALSE, show_check_vars = FALSE)
}
\arguments{
\item{lic_history}{data frame: output of \code{\link{make_lic_history}}}

\item{yrs}{numeric: years used for identifying lapse (note that the last year will
have all missing values for lapse output)}

\item{show_summary}{logical: if TRUE, include a tabular summary that shows lapse 
identification based on how many years until the next license is held.}

\item{show_check_vars}{logical: if TRUE, include output variables used in summary}
}
\description{
This is intended to be called following \code{\link{make_lic_history}} 
and codes lapse (0 = renews next year, 1 = lapses next year).
}
\examples{
library(dplyr)
data(history)

# excluding partial year 2019 since churn is only estimated for full-year results
history <- filter(history, year != 2019)

x <- select(history, -lapse) \%>\%
    identify_lapse(show_summary = TRUE) 

# calculate annual churn rate
# (note that a given year's churn is based on lapse summary from the previous year)
group_by(x, year) \%>\% 
    summarise(mean(lapse)) \%>\%
    mutate(year = year + 1)
}
\seealso{
Other license history functions: \code{\link{carry_duration}},
  \code{\link{carry_variables}},
  \code{\link{check_history_samp}},
  \code{\link{check_identify_R3}},
  \code{\link{check_identify_lapse}},
  \code{\link{identify_R3}},
  \code{\link{join_first_month}},
  \code{\link{make_lic_history}}, \code{\link{rank_sale}}
}
\concept{license history functions}
