% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{identify_lapse}
\alias{identify_lapse}
\title{Identify lapse group each year}
\usage{
identify_lapse(lic_history, yrs)
}
\arguments{
\item{lic_history}{license history data frame produced with \code{\link{make_lic_history}}}

\item{yrs}{numeric: Years in sales data (column 'year') from which
to create license history}
}
\description{
This is intended to be called following \code{\link{make_lic_history}} 
and codes lapse: 0 = renew next year, 1 = lapse next year.
}
\examples{
library(dplyr)
data(lic, sale)
sale_unranked <- left_join(sale, lic)
sale <- rank_sale(sale_unranked) \%>\%
    join_first_month(sale_unranked)
    
yrs <- 2008:2019
history <- sale \%>\%
    make_lic_history(yrs) \%>\%
    identify_lapse(yrs)
check_identify_lapse(history)
}
\seealso{
Other license history functions: \code{\link{check_history_samp}},
  \code{\link{check_identify_R3}},
  \code{\link{check_identify_lapse}},
  \code{\link{identify_R3}},
  \code{\link{join_first_month}},
  \code{\link{make_lic_history}}, \code{\link{rank_sale}}
}
\concept{license history functions}
