% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{identify_R3}
\alias{identify_R3}
\title{Identify R3 group each year}
\usage{
identify_R3(lic_history, yrs, show_summary = FALSE,
  show_check_vars = FALSE)
}
\arguments{
\item{lic_history}{license history data frame produced with \code{\link{make_lic_history}}}

\item{yrs}{numeric: Years in sales data (column 'year') from which
to create license history}

\item{show_summary}{logical: if TRUE, print a tabular summary that shows R3 category
by the number of years since a license was held.}

\item{show_check_vars}{logical: if TRUE, include output variables used in summary}
}
\description{
Intended to be called following \code{\link{make_lic_history}}, creates the
categorical variable, R3 (1=carried, 2=renewed, 3=reactivated, 4=recruited), 
where "retained" consists of carried + renewed (hence R3: retain, reactivate, recruit).
}
\examples{
library(dplyr)
data(history)

x <- history \%>\%
    select(-R3) \%>\%
    identify_R3(2008:2019, show_summary = TRUE)

# calculate \% of license holders who are recruits
group_by(x, year, R3) \%>\% 
    summarise(n = n()) \%>\% 
    mutate(pct = n / sum(n)) \%>\%
    filter(R3 == 4, year != 2019)
}
\seealso{
Other license history functions: \code{\link{check_history_samp}},
  \code{\link{check_identify_R3}},
  \code{\link{check_identify_lapse}},
  \code{\link{identify_lapse}},
  \code{\link{join_first_month}},
  \code{\link{make_lic_history}}, \code{\link{rank_sale}}
}
\concept{license history functions}
