% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{identify_R3}
\alias{identify_R3}
\title{Identify R3 group each year}
\usage{
identify_R3(lic_history, yrs)
}
\arguments{
\item{lic_history}{license history data frame produced with \code{\link{make_lic_history}}}

\item{yrs}{numeric: Years in sales data (column 'year') from which
to create license history}
}
\description{
This is intended to be called following \code{\link{make_lic_history}} and 
make the categorical variable R3 (1 = carried, 2 = renewed, 3 = reactivated, 4 = recruited), 
where "retained" consists of carried + renewed (hence R3: retain, reactivate, recruit).
}
\examples{
library(dplyr)
data(lic, sale)
sale_unranked <- left_join(sale, lic)
sale <- rank_sale(sale_unranked) \%>\%
    join_first_month(sale_unranked)
    
yrs <- 2008:2019
history <- sale \%>\%
    make_lic_history(yrs) \%>\%
    identify_R3(yrs)
check_identify_R3(history, yrs)

# calculate recruitment rate
group_by(history, year, R3) \%>\% 
summarise(n = n()) \%>\% 
    mutate(pct = n / sum(n)) \%>\%
    filter(R3 == 4, year != 2019)
}
\seealso{
Other license history functions: \code{\link{check_history_samp}},
  \code{\link{check_identify_R3}},
  \code{\link{check_identify_lapse}},
  \code{\link{identify_lapse}},
  \code{\link{join_first_month}},
  \code{\link{make_lic_history}}, \code{\link{rank_sale}}
}
\concept{license history functions}
