% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{identify_R3}
\alias{identify_R3}
\title{Identify R3 group each year}
\usage{
identify_R3(lic_history, yrs)
}
\arguments{
\item{lic_history}{license history data frame produced with \code{\link{make_lic_history}}}

\item{yrs}{numeric: Years in sales data (column 'year') from which
to create license history}
}
\description{
This is intended to be called following \code{\link{make_lic_history}} and 
codes R3: 1 = carried, 2 = renewed, 3 = reactivated, 4 = recruited, 
where "retained" consists of carried + renewed (hence R3: retain, reactivate, recruit).
}
\examples{
### Run ranking with sample data
library(tidyverse)
library(salic)
data(lic, sale, package = "salic")

sale_unranked <- select(lic, lic_id, duration) \%>\%
    right_join(sale) \%>\%
    select(cust_id, year, duration)
sale_ranked <- rank_sale(sale_unranked)

### Make license history
yrs <- 2004:2013
lic_history <- make_lic_history(sale_ranked, yrs)

### Identify R3
lic_history <- identify_R3(lic_history, yrs)

# check
check_identify_R3(lic_history, yrs)
}
\seealso{
Other license history functions: \code{\link{check_history_samp}},
  \code{\link{check_identify_R3}},
  \code{\link{check_identify_lapse}},
  \code{\link{identify_lapse}},
  \code{\link{join_first_month}},
  \code{\link{make_lic_history}}, \code{\link{rank_sale}}
}
\concept{license history functions}
