% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{lic}
\alias{lic}
\title{Sample Data: GA Hunting Licenses}
\format{A data frame with 32 rows and 4 variables
\describe{
\item{lic_id}{License ID}
\item{description}{License Description}
\item{priv}{Privilege Type}
\item{duration}{Duration (years, minimum of 1) for License
(where duration == 99 corresponds to a lifetime license)}
}}
\description{
A data frame describing hunting licenses
}
\seealso{
Other Sample Data: \code{\link{cust}}, \code{\link{sale}},
  \code{\link{state_abbreviations}}
}
\concept{Sample Data}
\keyword{datasets}
