% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{lic}
\alias{lic}
\title{Sample Data: License Types}
\format{A data frame with 105 rows and 4 variables
\describe{
\item{lic_id}{License ID (integer) (primary key)}
\item{type}{License Type (character) (hunt, fish, combo)}
\item{duration}{Duration in years (integer) (1=1yr/short-term, 2=2yr,...,99=lifetime)}
\item{lic_res}{License Residency (integer) (1=Res, 0=Nonres, NA=Unspecified)}
}}
\description{
Sample Data: License Types
}
\seealso{
Other Sample Data: \code{\link{cust}}, \code{\link{sale}},
  \code{\link{state_abbreviations}}
}
\concept{Sample Data}
\keyword{datasets}
