% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{make_history}
\alias{make_history}
\title{New license history function (in progress)}
\usage{
make_history(sale_ranked, yrs, carry_vars = NULL, calc_lapse = TRUE,
  show_diagnostics = FALSE)
}
\arguments{
\item{sale_ranked}{data frame: Sales table from which license history will be made}

\item{yrs}{numeric: Years in sales data (column 'year') from which
to create license history}

\item{carry_vars}{character: additional variables to carry over from previous year
(for multi-year and lifetime licenses).}

\item{calc_lapse}{logical: If TRUE, also estimate lapse (should only be done
for full-year time period)}
}
\description{
Get a faster function for making license history
}
\examples{
library(dplyr)
data(sale, lic)
sale_ranked <- left_join(sale, lic) \%>\% rank_sale()
history <- make_history(sale_ranked, 2008:2019)
}
\seealso{
Other license history functions: \code{\link{carry_duration}},
  \code{\link{carry_variables}},
  \code{\link{check_history_samp}},
  \code{\link{check_identify_R3}},
  \code{\link{check_identify_lapse}},
  \code{\link{identify_R3}}, \code{\link{identify_lapse}},
  \code{\link{join_first_month}},
  \code{\link{make_lic_history}}, \code{\link{prep_yrs}},
  \code{\link{rank_sale}}
}
\concept{license history functions}
