% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{make_history}
\alias{make_history}
\title{Make license history table}
\usage{
make_history(sale_ranked, yrs, carry_vars = NULL, yrs_lapse = yrs,
  show_diagnostics = FALSE)
}
\arguments{
\item{sale_ranked}{data frame: Sales table from which license history will be made; 
must include at least 3 variables (cust_id, year, duration)}

\item{yrs}{numeric: Years in sales data (column 'year') from which
to create license history}

\item{carry_vars}{character: additional variables to carry over from previous year
(for multi-year and lifetime licenses).}

\item{yrs_lapse}{numeric: years to include in lapse calculation (defaults to yrs). 
If NULL, lapse will not be calculated (useful for mid-year results)}

\item{show_diagnostics:}{If TRUE, will include intermediate variables in the
output dataset, necessary for running checks: \code{\link{history_check}}.}
}
\description{
The license history table accounts for multi-year/lifetime licenses by including
1 row for every year a license is held. The input data frame should only have 1 record 
per customer-year (ensured by running \code{\link{rank_sale}} beforehand).
}
\details{
The output data frame includes the following variables:
\itemize{
\item \emph{cust_id}: Customer ID
\item \emph{year}: License Year
\item \emph{carry_vars}: One or more optional variables to include; ensures 
multi-year/lifetime records are complete in future years.
\item \emph{duration_run}: A duration variable that accounts for multi-year/lifetimes.
\item \emph{lapse}: Lapse next year? (1=Yes, 0=No). Only included if yrs_lapse != NULL
\item \emph{R3}: R3 group this year (1=carried, 2=renewed, 3=reactivated, 4=recruited).  
Only included if more than 5 years are present.
}
Development Note: make_history() uses several \code{\link{history_internal}}
functions, all of which make use of package data.table for performance, 
see \href{https://github.com/Rdatatable/data.table/wiki}{data.table wiki} &
\href{https://github.com/Rdatatable/data.table/wiki/Getting-started}{getting started vignette}
for more information.
}
\examples{
library(dplyr)
data(sale, lic)
sale_ranked <- left_join(sale, lic) \%>\% rank_sale()
history <- make_history(sale_ranked, 2008:2018, "res")

# history includes more rows than sale_ranked if multi-year/lifetimes are present
left_join(count(history, year), count(sale_ranked, year), by = "year")
}
\seealso{
Other license history functions: \code{\link{history_check}},
  \code{\link{rank_sale}}
}
\concept{license history functions}
