% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{make_lic_history}
\alias{make_lic_history}
\title{Make a License History Table}
\usage{
make_lic_history(sale_ranked, yrs, carry_vars = NULL)
}
\arguments{
\item{sale_ranked}{data frame: Sales table from which license history will be made}

\item{yrs}{numeric: Years in sales data (column 'year') from which
to create license history}

\item{carry_vars}{character: additional variables to carry over from previous year
(for multi-year and lifetime licenses).}
}
\description{
Output table is stored in a data frame with 1 row per customer-year
}
\details{
The input data frame (sale) must have at least 3 columns
(cust_id, year, duration) and should only have 1 record per customer per year
(which can be insured by running \code{\link{rank_sale}}).

Each data frame in the output data frame has at least 6 columns:
\itemize{
\item \emph{cust_id}: Customer ID
\item \emph{year}: Privilege year
\item \emph{duration}: Bought a license this year? (>=1: yes, ==0: no),
where duration = 1 refers to annual or short-term, 2 to 2 yr, 3 to 3 yr, ...,
duration = 99 is used to indicate lifetime licenses
\item \emph{duration_run}: Running duration, which is necessary for multi-year
and lifetime licenses
\item \emph{lag_duration_run}: Previous year duration_run. A temporary value
used in coding and for downstream checking
\item \emph{bought}: Bought a license this year? (TRUE, FALSE). If FALSE, that 
indicates a license carried over from a previous year
}
}
\examples{
### Run ranking with sample data
library(tidyverse)
library(salic)
data(lic, sale, package = "salic")

sale_unranked <- select(lic, lic_id, duration) \%>\%
    right_join(sale) \%>\%
    select(cust_id, year, duration)
sale_ranked <- rank_sale(sale_unranked)

### Make license history
yrs <- 2004:2013
lic_history <- make_lic_history(sale_ranked, yrs)

# check
check_history_samp(lic_history)
}
\seealso{
Other license history functions: \code{\link{check_history_samp}},
  \code{\link{check_identify_R3}},
  \code{\link{check_identify_lapse}},
  \code{\link{identify_R3}}, \code{\link{identify_lapse}},
  \code{\link{rank_sale}}
}
