% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history-deprecated.R
\name{make_priv}
\alias{make_priv}
\title{Make a Privilege Table from License History}
\usage{
make_priv(track)
}
\arguments{
\item{track}{list: Set of license history tables created using
\code{\link{make_track}}}
}
\description{
Tags churn/renewal and R3 categories at the customer scale over a set of years.
This allows efficient summarization of these metrics at any aggregate scale.
}
\details{
Each data frame in the output list has 3 columns:
\itemize{
\item \emph{cust_id}: Customer ID
\item \emph{lapse}: Lapsed next year? (==1: yes, ==0: no)
\item \emph{R3}: R3 group this year (==0: carry, ==1: retain, ==2: reactivate, ==3: recruit)
}
}
\examples{
make_priv()
}
\seealso{
Other deprecated license history functions: \code{\link{check_R3}},
  \code{\link{check_lapse}}, \code{\link{check_make_priv}},
  \code{\link{check_make_track}},
  \code{\link{check_track}}, \code{\link{get_R3}},
  \code{\link{get_lapse}}, \code{\link{make_priv_final}},
  \code{\link{make_track}}, \code{\link{update_track}}
}

