% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{new_dashboard}
\alias{new_dashboard}
\title{Setup a new state dashboard with default directories and template scripts}
\usage{
new_dashboard(state, time_period, sa_path = "E:/SA",
  R_version = "3.4.3", project_library = "data-dashboards")
}
\arguments{
\item{state}{character: Two letter state designation}

\item{time_period}{character: Time period for the first dashboard 
(e.g., "2015", "2016-q1", etc.)}

\item{sa_path}{character: File path to the Southwick main folder 
(for analysis and data, etc.)}

\item{R_version}{character: Version of R to use for this project}

\item{project_library}{character: Name of project-specific R package library}
}
\description{
This is intended to be run before data processing for a new state begins
}
\examples{
salic::new_dashboard("XX", "2017")
}
\concept{functions for making directories and files}
