% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{pct_round}
\alias{pct_round}
\title{Round numeric values and print as percent (internal salic only - not exported)}
\usage{
pct_round(x, rnd = 1, scale = 100)
}
\arguments{
\item{x}{numeric: Vector of values to display as rounded percentages}

\item{rnd}{numeric: number of decimals to round pct change results}

\item{scale}{numeric: scaling paramter - defaults to 100 for showing percentages}
}
\description{
This is a helper function for conveniently displaying percentages. It is used
in other summary functions in salic.
}
\examples{
x <- data.frame(id = c(1,2,3,4), pop = c(135, 416, 389, 320))
x$pct <- pct_round(x$pop / sum(x$pop))
}
\seealso{
Other internal helper functions: \code{\link{calc_churn}},
  \code{\link{format_num}}, \code{\link{print_dat}}
}
\concept{internal helper functions}
\keyword{internal}
