% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{print_dat}
\alias{print_dat}
\title{Print a data frame with caption/note (internal salic only - not exported)}
\usage{
print_dat(x, caption = NULL, note = NULL, digits = 1,
  big.mark = ",")
}
\arguments{
\item{x}{data frame: data frame contents to print}

\item{caption}{character: Optional caption to print}

\item{note}{character: Optional note(s) to print.
for multiple lines of notes}

\item{digits}{number of digits for rounding}

\item{big.mark}{character: separator between 1000s}
}
\description{
Intended for showing tables with titles & notes in logged output in doc/
}
\examples{
x <- data.frame(yr = c(2005, 2006), cust = c(100000, 131000),
    sales = c(567891, 673568), churn = c(NA, 25.23), char = c("test", NA))
print_dat(x)
print_dat(x, "Customer Sales by Year")
print_dat(x, "Customer Sales by Year", "A note!")
print_dat(x, "Customer Sales by Year", big.mark = "")
print_dat(x, "Customer Sales by Year", digits = 0)
}
\seealso{
Other internal helper functions: \code{\link{calc_churn}},
  \code{\link{format_num}}, \code{\link{pct_round}}
}
\concept{internal helper functions}
\keyword{internal}
