% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{rank_sale}
\alias{rank_sale}
\title{Filter sales to 1 row per customer per year.}
\usage{
rank_sale(sale, rank_var = "duration", grp_var = c("cust_id", "year"),
  first_month = FALSE)
}
\arguments{
\item{sale}{data frame: Input sales data}

\item{rank_var}{character: name of variable(s) to use for ranking}

\item{grp_var}{character: name of variable(s) used for grouping}

\item{first_month}{logical: If TRUE, also runs \code{\link{join_first_month}}
to ensure the output contains the earliest month by grp_var 
(useful for mid-year dashboards)}
}
\description{
Intended to run before \code{\link{make_lic_history}} to ensure that 
multi-year/lifetime sales are accounted for.  
The default (and intended purpose) is to pick the maximum "duration" value 
per customer-year. Optionally, it will also pick the minimum value of month 
(useful in mid-year dashboards) if first_month = TRUE.
}
\examples{
library(dplyr)
data(lic, sale)

sale_unranked <- left_join(sale, lic)
sale_ranked <- rank_sale(sale_unranked)
    
# check sale ranking - highest duration will always be picked
left_join(
    count(sale_ranked, duration), 
    distinct(sale_unranked, cust_id, year, duration) \%>\% count(duration), 
    by = "duration",
    suffix = c(".ranked", ".unranked")
)

# with earliest month included
sale_ranked <- rank_sale(sale_unranked, first_month = TRUE)
left_join(
    count(sale_ranked, month), 
    distinct(sale_unranked, cust_id, year, month) \%>\% count(month), 
    by = "month",
    suffix = c(".ranked", ".unranked")
)
}
\seealso{
Other license history functions: \code{\link{carry_duration}},
  \code{\link{carry_variables}},
  \code{\link{check_history_samp}},
  \code{\link{check_identify_R3}},
  \code{\link{check_identify_lapse}},
  \code{\link{identify_R3}}, \code{\link{identify_lapse}},
  \code{\link{join_first_month}},
  \code{\link{make_lic_history}}, \code{\link{prep_yrs}}
}
\concept{license history functions}
