% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{rank_sale}
\alias{rank_sale}
\title{Aggregate a Sale Table to insure 1 row per customer per year.}
\usage{
rank_sale(sale, rank_var = "duration")
}
\arguments{
\item{sale}{data frame: Input sales data}

\item{rank_var}{character: name of variable(s) to use for ranking}
}
\description{
This function takes an input sale table and returns the max duration
value present in each year (per customer) - essentially a ranking by
license duration.
}
\examples{
### Run ranking with sample data
library(dplyr)
library(tidyr)

# Join duration variable to sale data
sale_unranked <- select(lic, lic_id, duration) \%>\%
    right_join(sale) \%>\%
    select(cust_id, year, duration)

# Perform ranking and check the result
sale_ranked <- rank_sale(sale_unranked)
check_rank_sale(sale_ranked, sale_unranked)
}
\seealso{
\code{\link{check_rank_sale}}

Other functions for summarizing license history: \code{\link{get_R3}},
  \code{\link{get_lapse}}, \code{\link{make_priv_final}},
  \code{\link{make_priv}}, \code{\link{make_track}},
  \code{\link{update_track}}
}

