% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license-history.R
\name{rank_sale}
\alias{rank_sale}
\title{Aggregate a Sale Table to insure 1 row per customer per year.}
\usage{
rank_sale(sale, rank_var = "duration", grp_var = c("cust_id", "year"))
}
\arguments{
\item{sale}{data frame: Input sales data}

\item{rank_var}{character: name of variable(s) to use for ranking}

\item{grp_var}{character: name of variable(s) used for grouping}
}
\description{
This function takes an input sale table and returns the max duration
value present in each year (per customer) - essentially a ranking by
license duration.
}
\examples{
### Run ranking with sample data
library(tidyverse)
data(lic, sale, package = "salic")

# Join duration variable to sale data
sale_unranked <- select(lic, lic_id, duration) \%>\%
    right_join(sale) \%>\%
    select(cust_id, year, duration)

# Perform ranking and check the result
sale_ranked <- rank_sale(sale_unranked)
}
\seealso{
\code{\link{check_rank_sale}}

Other license history functions: \code{\link{check_history_samp}},
  \code{\link{check_identify_R3}},
  \code{\link{check_identify_lapse}},
  \code{\link{identify_R3}}, \code{\link{identify_lapse}},
  \code{\link{make_lic_history}}
}
\concept{license history functions}
