% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category-variables.R, R/dashboard.R
\name{recode_agecat}
\alias{recode_agecat}
\title{Create age category based on year (of sale) and birth year}
\usage{
recode_agecat(dat, age_labs = c("0-17", "18-24", "25-34", "35-44",
  "45-54", "55-64", "65+"), age_breaks = c(-Inf, 17, 24, 34, 44, 54, 64,
  Inf), max_age = 110, suppress_check = TRUE)

recode_agecat(dat, age_labs = c("0-17", "18-24", "25-34", "35-44",
  "45-54", "55-64", "65+"), age_breaks = c(-Inf, 17, 24, 34, 44, 54, 64,
  Inf), max_age = 110, suppress_check = TRUE)
}
\arguments{
\item{dat}{data frame: Input table}

\item{age_labs}{character: labels to use for age category}

\item{age_breaks}{numeric: breaks for age category passed to 
\code{\link[base]{cut}}}

\item{max_age}{numeric: maximum allowed age. Anything above will be set to missing.}

\item{suppress_check}{logical: If TRUE, does not print a coding summary}

\item{dat}{data frame: Input table}

\item{age_labs}{character: labels to use for age category}

\item{age_breaks}{numeric: breaks for age category passed to 
\code{\link[base]{cut}}}

\item{max_age}{numeric: maximum allowed age. Anything above will be set to missing.}

\item{suppress_check}{logical: If TRUE, does not print a coding summary}
}
\description{
Uses 'birth_year' (from cust) and 'year' (from sale) to identify
age for each license year

Uses 'birth_year' (from cust) and 'year' (from sale) to identify
age for each license year
}
\examples{
library(dplyr)
data(history)
x <- recode_agecat(history, suppress_check = FALSE)
library(dplyr)
data(cust, sale)
sale2 <- left_join(sale, cust) \%>\% 
    recode_agecat()
}
\concept{functions for recoding license data}
