% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{recode_month}
\alias{recode_month}
\title{Standardize month (for individual state dashboards)}
\usage{
recode_month(sale, month_range = 0:12)
}
\arguments{
\item{sale}{data frame: Input sales table with at lest 2 variables: dot
(transaction date in 'yyyy-mm-dd') and year (numeric license year)}

\item{month_range}{numeric: A vector of months allowed in the output. Defaults
to 0:12 since this is a common range of sales. Any months outside the range will be
set to either the lowest values (for those less than the range) or the
highest value (for those over). This prevents unusual sale months from appearing
in results.}
}
\value{
Returns a sales table with a standardized 'month' variable and
prints a validation output.
}
\description{
DON'T USE FOR NAT/REG PREPARATION (where year should be based on sale date)
}
\details{
The output month variable captures transaction year and month in one metric.
The month number is relative to the license year: 1 = Jan, 2 = Feb, etc, for
current/future years and 0 = Dec, -1 = Nov, ..., for previous years.
}
\examples{
# For Southwick analysts: see data preparation scripts
}
\seealso{
Other functions for standardizing state license data: \code{\link{recode_state}}
}
\concept{functions for standardizing state license data}
