% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_state}
\alias{recode_state}
\title{Recode a state variable to deal with unusual values}
\usage{
recode_state(dat, state_table, oldvar = "state", newvar = "state_new")
}
\arguments{
\item{dat}{data frame: input data with a state variable}

\item{state_table}{data frame: input data holding valid abbreviations}

\item{oldvar}{character: name of state variable}

\item{newvar}{character: name of new state variable}
}
\description{
The idea is to set any odd values to missing. It's a simple function that does
some minimal recoding
}
\examples{
library(salic)
data(state_abbreviations)
data(cust)
x <- recode_state(cust, state_abbreviations)
dplyr::count(x, state_new, state)
}
\seealso{
Other functions for recoding license data: \code{\link{recode_agecat}},
  \code{\link{recode_month}}
}
