% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{sale}
\alias{sale}
\title{Sample Data: GA Hunting Sales}
\format{A data frame with 2802 rows and 4 variables
\describe{
\item{cust_id}{Customer ID}
\item{lic_id}{License ID}
\item{dot}{Date of Transaction}
\item{year}{License Year}
}}
\description{
A data frame with GA Hunting sales corresponding to the 1000 customer sample
}
\seealso{
Other Sample Data: \code{\link{cust}}, \code{\link{lic}},
  \code{\link{state_abbreviations}}
}
\concept{Sample Data}
\keyword{datasets}
