% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{sale}
\alias{sale}
\title{Sample Data: Sales (Jan 1, 2008 through June 30, 2019)}
\format{A data frame with 5 variables
\describe{
\item{cust_id}{Customer ID (integer) (foreign key)}
\item{lic_id}{License ID (integer) (foreign key)}
\item{year}{License Year (integer) (yyyy)}
\item{month}{Transaction Purchase Month (integer) (1=Jan, 2=Feb, ..., 12=Dec)}
\item{res}{Residency (integer) (1=Res, 0=Nonres)}
}}
\description{
Sample Data: Sales (Jan 1, 2008 through June 30, 2019)
}
\seealso{
Other Sample Data: \code{\link{cust}},
  \code{\link{history}}, \code{\link{lic}},
  \code{\link{metrics}}, \code{\link{state_abbreviations}}
}
\concept{Sample Data}
\keyword{datasets}
