% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{sale}
\alias{sale}
\title{Sample Data: Sales 2008 thru 2018}
\format{A data frame with 5 variables
\describe{
\item{cust_id}{Customer ID (foreign key)}
\item{lic_id}{License ID (foreign key)}
\item{year}{License Year (numeric) (yyyy)}
\item{month}{Transaction Purchase Month (numeric) (1 = Jan, 2 = Feb, ...)}
\item{res}{Residency (numeric) (1 = State Resident, 0 = State Nonresident, 
NA = Unknown)}
}}
\description{
Sample Data: Sales 2008 thru 2018
}
\seealso{
Salic Function Reference: \code{\link{salic}}

Other Sample Data: \code{\link{cust}},
  \code{\link{dashboard}}, \code{\link{data_salic}},
  \code{\link{history}}, \code{\link{lic}},
  \code{\link{metrics}}, \code{\link{state_abbreviations}}
}
\concept{Sample Data}
\keyword{datasets}
