% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{state_abbreviations}
\alias{state_abbreviations}
\title{Data: State/Province/Territory Abbreviations for US & Canada}
\format{A data frame with 72 rows and 3 variables
\describe{
\item{name}{State/Province/Territory name}
\item{state}{Abbreviation}
\item{country}{Country Abbreviation where CN = Canada}
}}
\description{
A data frame with 72 abbreviations, used in the \code{\link{recode_state}} function
}
\seealso{
Other Sample Data: \code{\link{cust}},
  \code{\link{history}}, \code{\link{lic}},
  \code{\link{metrics}}, \code{\link{sale}}
}
\concept{Sample Data}
\keyword{datasets}
