% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{summary_churn}
\alias{summary_churn}
\title{Summarize churn by year}
\usage{
summary_churn(x, years, rnd = 1, out = NULL, title = "Churn by Year",
  note = NULL, suppress_notes = FALSE)
}
\arguments{
\item{x}{data frame: Table holding sales by year with a minimum of
2 variables (cust_id, year)}

\item{years}{numeric: range of years needed to calculate churn}

\item{rnd}{numeric: number of decimals to round pct change results}

\item{out}{character: file path to optional output csv}

\item{title}{character: title for output table}

\item{note}{character: note for output table}

\item{suppress_notes}{logical: If TRUE, returns a data frame only}
}
\description{
This is a simple calculation that doesn't take into account multi-year
licenses, so it should only be used for initial data validation
}
\examples{
library(dplyr)
data(sale, lic)
summary_churn(sale, 2012:2018)

sale <- left_join(sale, lic)
filter(sale, type \%in\% c("hunt", "combo")) \%>\%
    summary_churn(2012:2018, title = "Hunting Churn", note = "A note")
}
\seealso{
Other functions for validating license data: \code{\link{check_dups}},
  \code{\link{count_lines_textfile}},
  \code{\link{summary_initial}}, \code{\link{summary_sale}}
}
\concept{functions for validating license data}
