% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{summary_sale}
\alias{summary_sale}
\title{Summarize sales and customers by year}
\usage{
summary_sale(x, include_revenue = FALSE, rnd = 1, out = NULL,
  title = "Summarize Sales & Customers", note = NULL,
  suppress_notes = FALSE)
}
\arguments{
\item{x}{data frame: Table holding sales by year with a minimum of
2 variables (cust_id, year)}

\item{include_revenue}{logical: If TRUE, revenue by year will also be shown}

\item{rnd}{numeric: number of decimals to round pct change results}

\item{out}{character: file path to optional output csv}

\item{title}{character: title for output table}

\item{note}{character: note for output table}

\item{suppress_notes}{logical: If TRUE, returns a data frame only}
}
\description{
Summarize sales and customers by year
}
\examples{
library(dplyr)
data(sale)
summary_sale(sale)

# example with made-up license revenue
sale2 <- mutate(sale, revenue = 30)
summary_sale(sale2, rnd = 2, include_revenue = TRUE)
}
\seealso{
Other functions for validating license data: \code{\link{check_dups}},
  \code{\link{count_lines_textfile}},
  \code{\link{summary_churn}},
  \code{\link{summary_initial}}
}
\concept{functions for validating license data}
